@@ -17,14 +17,19 @@ import logging
 import sys
 from dataclasses import asdict, field
 from pathlib import Path
-from typing import Dict, List, Optional
 
 import hydra
 from tqdm import tqdm
 
+from nemo_skills.code_execution import extract_error_message
+from nemo_skills.code_execution.math_grader import extract_answer
 from nemo_skills.code_execution.sandbox import get_sandbox, sandbox_params
 from nemo_skills.inference.prompt.utils import Prompt, PromptConfig, datasets, prompt_types
-from nemo_skills.inference.server.model import ErrorRecoveryConfig, get_model, server_params
+from nemo_skills.inference.server.code_execution_model import (
+    ErrorRecoveryConfig,
+    get_code_execution_model,
+    server_params,
+)
 from nemo_skills.utils import get_fields_docstring, get_help_message, nested_dataclass, setup_logging
 
 LOG = logging.getLogger(__file__)
@@ -56,11 +61,9 @@ class GenerateSolutionsConfig:
 
     # Can specify one of the existing datasets.
     # Choices: {datasets}.
-    dataset: Optional[str] = None
-    split_name: Optional[str] = None  # Can be train, validation, test or train_full (train + validation)
-    data_file: Optional[str] = None  # Can directly specify a data file, if using a custom dataset
-
-    example_dicts: Optional[List[Dict]] = None  # A list of few-shot demonstrations to be shown in the prompt
+    dataset: str | None = None
+    split_name: str | None = None  # Can be train, validation, test or train_full (train + validation)
+    data_file: str | None = None  # Can directly specify a data file, if using a custom dataset
 
     batch_size: int = 16
     max_samples: int = -1  # If > 0, will stop after generating this many samples. Useful for debugging
@@ -85,13 +88,19 @@ cs = hydra.core.config_store.ConfigStore.instance()
 cs.store(name="base_generation_config", node=GenerateSolutionsConfig)
 
 
+def add_answer_and_error_message(output: dict):
+    output['predicted_answer'] = extract_answer(output['generation'])
+    if 'error_message' not in output:
+        output['error_message'] = extract_error_message(output['generation'])
+
+
 @hydra.main(version_base=None, config_name='generation_config', config_path='.')
 def generate_solutions(cfg: GenerateSolutionsConfig):
     cfg = GenerateSolutionsConfig(_init_nested=True, **cfg)
 
     LOG.info("Config used: %s", cfg)
     sandbox = get_sandbox(**cfg.sandbox) if cfg.sandbox is not None else None
-    llm = get_model(**cfg.server, sandbox=sandbox)
+    llm = get_code_execution_model(**cfg.server, sandbox=sandbox)
 
     # making sure output folder exists
     Path(cfg.output_file).absolute().parent.mkdir(parents=True, exist_ok=True)
@@ -115,37 +124,47 @@ def generate_solutions(cfg: GenerateSolutionsConfig):
 
     # additionally, skipping whatever is pre-filled, assuming offset didn't change
     data = data[starting_idx:]
+    prompt = Prompt(config=cfg.prompt)
+
+    if cfg.max_samples < 0:
+        cfg.max_samples = len(data)
 
     # setting buffering=1 to force to dump the output after every line, so that we can see intermediate generations
     with open(cfg.output_file, "at" if cfg.skip_filled else "wt", encoding="utf-8", buffering=1) as fout:
-        prompts = []
         data_points = []
         for idx, data_point in tqdm(enumerate(data), initial=starting_idx, total=len(data) + starting_idx):
-            if idx == cfg.max_samples:
+            if idx >= cfg.max_samples:
                 break
-
-            prompts.append(Prompt(config=cfg.prompt, input_dict=data_point, example_dicts=cfg.example_dicts))
 
             data_points.append(data_point)
 
-            if len(prompts) == cfg.batch_size:
+            if len(data_points) == cfg.batch_size:
                 # batch-computing the outputs
-
-                outputs = llm(stop_phrases=list(cfg.prompt.stop_phrases), prompts=prompts, **asdict(cfg.inference))
+                outputs = llm.generate(
+                    prompts=[prompt.build_string(data_point) for data_point in data_points],
+                    stop_phrases=list(cfg.prompt.stop_phrases),
+                    **asdict(cfg.inference),
+                )
 
                 for output, original_data_point in zip(outputs, data_points):
                     # to make it easier to follow up with evaluation and limit accidental errors, we are adding
                     # all of the ground-truth data to the output file alongside the generated solutions
                     output.update(original_data_point)
+                    # adding answer and error message
+                    add_answer_and_error_message(output)
                     fout.write(json.dumps(output) + "\n")
-                prompts = []
                 data_points = []
 
         # collecting the final batch
-        if len(prompts) > 0:
-            outputs = llm(stop_phrases=list(cfg.prompt.stop_phrases), prompts=prompts, **asdict(cfg.inference))
+        if len(data_points) > 0:
+            outputs = llm.generate(
+                prompts=[prompt.build_string(data_point) for data_point in data_points],
+                stop_phrases=list(cfg.prompt.stop_phrases),
+                **asdict(cfg.inference),
+            )
             for output, original_data_point in zip(outputs, data_points):
                 output.update(original_data_point)
+                add_answer_and_error_message(output)
                 fout.write(json.dumps(output) + "\n")
 
 
