@@ -18,49 +18,35 @@ import json
 import logging
 import os
 import re
-from collections import Counter
-from concurrent.futures import ThreadPoolExecutor
-from typing import List, Union
+import time
+from typing import Union
 
 import requests
 
-from nemo_skills.code_execution import (
-    CODE_OUTPUT_SEPARATORS,
-    CODE_SEPARATORS,
-    extract_code_to_execute,
-    extract_error_message,
-)
-from nemo_skills.code_execution.math_grader import extract_answer
-from nemo_skills.code_execution.sandbox import Sandbox
-from nemo_skills.inference.prompt.utils import Prompt
-from nemo_skills.utils import nested_dataclass, python_doc_to_cmd_help
-
 LOG = logging.getLogger(__name__)
 
-# TODO: we should make this more efficient by asynchronously submitting eval
-#       requests and getting new prompts in a queue to make sure we can always
-#       pack full batches, instead of only resubmitting a small set of prompts
-#       that require additional code executions
+
+def remove_stop_phrases(text: str, stop_phrases: list[str]) -> str:
+    """Removes everything after the last stop token."""
+    if not stop_phrases:
+        return text
+    return re.split("|".join([sp.replace('|', '\\|') for sp in stop_phrases]), text, maxsplit=1)[0]
 
 
-@nested_dataclass
-class ErrorRecoveryConfig:
-    # Number of attempts to recover from code execution error
-    recovery_attempts: int = 0
-    # If true, take code block based on majority voting of `recovery_attempts` code outputs.
-    # Otherwise take the first valid code output.
-    # So `majority_voting=False` is potentially faster.
-    majority_voting: bool = True
-    # Temperature for recovery requests
-    temperature: float = 0.7
-    # Top-p for recovery requests
-    top_p: float = 0.95
-    # Top-k for recovery requests
-    top_k: int = 0
+def preprocess_request(request: dict):
+    """Just a small utility to pre-process some of the parameters of request."""
+    # temperature of 0 means greedy, but it's not always supported by the server
+    # so setting explicit greedy parameters instead
+    if request["temperature"] == 0:
+        request["temperature"] = 1.0
+        request["top_k"] = 1
+        request["top_p"] = 1.0
 
 
-def remove_stop_tokens(text: str, stop_phrases: List[str]) -> str:
-    return re.split("|".join(stop_phrases), text, maxsplit=1)[0]
+def postprocess_output(outputs: list[dict], stop_phrases: list[str]):
+    """Post-processes the outputs of the model."""
+    for output in outputs:
+        output['generation'] = remove_stop_phrases(output['generation'], stop_phrases)
 
 
 class BaseModel(abc.ABC):
@@ -69,287 +55,72 @@ class BaseModel(abc.ABC):
     Args:
         host: Optional[str] = '127.0.0.1' - Host of the inference server.
         port: Optional[str] = '5000' - Port of the inference server.
-        sandbox: Optional[Sandbox] = None - Sandbox for executing code.
             Only required if handle_code_execution is True.
         ssh_server: Optional[str] = None - SSH server for tunneling requests.
             Useful if server is running on slurm cluster to which there is an ssh access
             Can also be specified through NEMO_SKILLS_SSH_SERVER env var.
         ssh_key_path: Optional[str] = None - Path to the ssh key for tunneling.
             Can also be specified through NEMO_SKILLS_SSH_KEY_PATH env var.
-        max_code_output_characters: Optional[int] = 1000 - Maximum number of characters for code execution output.
-        code_execution_timeout: Optional[float] = 10.0 - Timeout for code execution in seconds.
-        max_code_executions: Optional[int] = 3 - Maximum number of code executions per generation.
-        stop_on_code_error: Optional[bool] = True - Whether to stop generation if code execution fails.
-        handle_code_execution: Optional[bool] = True - Whether to handle code execution in this class
-            or make a single call to the server. If set to False, the server needs to have special logic
-            for communicating with the sandbox.
-        error_recovery: Optional[dict] = None - Configuration for error recovery.
     """
 
     def __init__(
         self,
-        host='127.0.0.1',
-        port='5000',
-        sandbox=None,
-        ssh_server=None,
-        ssh_key_path=None,
-        max_code_output_characters=1000,
-        code_execution_timeout=10.0,
-        max_code_executions=3,
-        stop_on_code_error=True,
-        handle_code_execution=True,
-        error_recovery=None,
+        host: str = '127.0.0.1',
+        port: str = '5000',
+        ssh_server: str | None = None,
+        ssh_key_path: str | None = None,
     ):
         self.server_host = host
         self.server_port = port
         self.ssh_server = os.getenv("NEMO_SKILLS_SSH_SERVER", ssh_server)
         self.ssh_key_path = os.getenv("NEMO_SKILLS_SSH_KEY_PATH", ssh_key_path)
-        self.max_code_output_characters = max_code_output_characters
-        self.code_execution_timeout = code_execution_timeout
-        self.max_code_executions = max_code_executions
-        if error_recovery is None:
-            error_recovery = {}
-        self.error_recovery = ErrorRecoveryConfig(_init_nested=True, **error_recovery)
-        self.handle_code_execution = handle_code_execution
-        self.stop_on_code_error = stop_on_code_error
-        if self.handle_code_execution and sandbox is None:
-            raise ValueError("Sandbox is required for handling code execution")
-        if not self.handle_code_execution and not stop_on_code_error:
-            # TODO: same warning for other ignored parameters
-            LOG.warning("When code execution is not handled here, stop_on_code_error is ignored.")
-        self.sandbox = sandbox
-
-    @abc.abstractmethod
-    def _single_call(
-        self,
-        prompts,
-        tokens_to_generate,
-        temperature,
-        top_p,
-        top_k,
-        repetition_penalty,
-        random_seed,
-        stop_phrases: List[str],
-    ):
-        pass
-
-    def __call__(
-        self,
-        prompts,
-        tokens_to_generate,
-        temperature,
-        top_p,
-        top_k,
-        repetition_penalty,
-        random_seed,
-        stop_phrases: List[str],
-    ):
-        if self.handle_code_execution:
-            full_stop_phrases = stop_phrases + [CODE_SEPARATORS[-1]]
-        else:
-            full_stop_phrases = stop_phrases
-        # prompts are added later
-        request = {
-            "tokens_to_generate": tokens_to_generate,
-            "temperature": temperature,
-            "top_k": top_k,
-            "top_p": top_p,
-            "random_seed": random_seed,
-            "repetition_penalty": repetition_penalty,
-            "stop_phrases": full_stop_phrases,
-        }
-
-        # if code execution is handled by the inference framework, we only need to make a single call
-        # and then apply postprocessing to extract errors from the output
-        if not self.handle_code_execution:
-            request["prompts"] = prompts
-            outputs = self._single_call(**request)
-            outputs = [
-                {
-                    'generated_solution': remove_stop_tokens(output, stop_phrases),
-                    'predicted_answer': extract_answer(output),
-                    'error_message': extract_error_message(output),
-                }
-                for output in outputs
-            ]
-            return outputs
-
-        # making requests to LLM and iterating on prompts that produce code tokens
-        # after executing code and getting result back into the prompt
-        new_outputs = [
-            {
-                'full_prompt': prompts[idx],
-                'result': None,
-                'error_message': Sandbox.NOT_EXECUTED,
-                'session_id': None,
-            }
-            for idx in range(len(prompts))
-        ]
-        remaining_ids = list(range(len(new_outputs)))
-        num_executions = 0
-        with ThreadPoolExecutor(max_workers=len(prompts)) as executor:
-            while len(remaining_ids) > 0:
-                num_executions += 1
-                request["prompts"] = [new_outputs[idx]['full_prompt'] for idx in remaining_ids]
-
-                outputs = self._single_call(**request)
-                new_ids = []
-                # checking if any of the outputs need code execution and submitting requests in parallel
-                futures = [None] * len(prompts)
-                for idx, output in zip(remaining_ids, outputs):
-                    if output.endswith(CODE_SEPARATORS[-1]):
-                        futures[idx] = executor.submit(
-                            self.sandbox.execute_code,
-                            generated_code=extract_code_to_execute(output),
-                            timeout=self.code_execution_timeout,
-                            max_output_characters=self.max_code_output_characters,
-                            session_id=new_outputs[idx]['session_id'],
-                        )
-                for idx, output in zip(remaining_ids, outputs):
-                    if output.endswith(CODE_SEPARATORS[-1]):
-                        result, new_outputs[idx]['session_id'] = futures[idx].result()
-                        # for now if there is any error or no output, we stop generation
-                        # might revise in the future to allow LLM to recover
-                        if result['error_message']:
-                            new_outputs[idx]['error_message'] = result['error_message']
-                            if self.stop_on_code_error:
-                                new_outputs[idx]['full_prompt'].generated_solution += output
-                                continue
-                            text_only_part = output.split(CODE_SEPARATORS[0])[0]
-                            new_outputs[idx]['full_prompt'].generated_solution += text_only_part
-                            code_output = self._recover_from_error(request, new_outputs[idx], executor)
-                            # if re-generation did not help
-                            if code_output is None:
-                                code_output = result["result"]
-                                new_outputs[idx]['full_prompt'].generated_solution += output[len(text_only_part) :]
-                        else:
-                            new_outputs[idx]['full_prompt'].generated_solution += output
-                            new_outputs[idx]['error_message'] = ''
-                            code_output = result["result"]
-
-                        # adding code output to the prompt
-                        code_output = f'\n{CODE_OUTPUT_SEPARATORS[0]}\n{code_output}\n{CODE_OUTPUT_SEPARATORS[1]}\n'
-                        new_outputs[idx]['full_prompt'].generated_solution += code_output
-                        # setting a limit on max code executions to speed things up
-                        # (sometimes keeps repeating the same sequence forever)
-                        if num_executions >= self.max_code_executions:
-                            new_outputs[idx]['error_message'] = "Max code executions reached"
-                        else:
-                            new_ids.append(idx)
-                    else:
-                        new_outputs[idx]['full_prompt'].generated_solution += output
-                remaining_ids = new_ids
-
-        # removing original prompt and stop tokens from the end of the generated text
-        outputs = []
-        for output in new_outputs:
-            if output['session_id'] is not None:
-                self.sandbox.clear_session(output['session_id'])
-            generated_solution = remove_stop_tokens(output['full_prompt'].generated_solution, stop_phrases)
-            outputs.append(
-                {
-                    'generated_solution': generated_solution,
-                    'predicted_answer': extract_answer(generated_solution),
-                    'error_message': output['error_message'],
-                }
-            )
-        return outputs
-
-    def _recover_from_error(self, request, new_output, executor):
-        recovery_request = {key: value for key, value in request.items() if key != 'prompts'}
-        recovery_request['prompts'] = [new_output['full_prompt']]
-
-        recovery_request['temperature'] = self.error_recovery.temperature
-        recovery_request['top_p'] = self.error_recovery.top_p
-        recovery_request['top_k'] = self.error_recovery.top_k
-
-        outputs = []
-        futures = [None] * self.error_recovery.recovery_attempts
-        results = [None] * self.error_recovery.recovery_attempts
-        for rs in range(self.error_recovery.recovery_attempts):
-            recovery_request['random_seed'] = rs
-            output = self._single_call(**recovery_request)[0]
-            outputs.append(output)
-            if output.endswith(CODE_SEPARATORS[-1]):
-                futures[rs] = executor.submit(
-                    self.sandbox.execute_code,
-                    generated_code=extract_code_to_execute(output),
-                    timeout=self.code_execution_timeout,
-                    max_output_characters=self.max_code_output_characters,
-                    session_id=new_output['session_id'],
-                )
-
-                if not self.error_recovery.majority_voting:
-                    result, _ = futures[rs].result()
-                    # quit on first correct output if not majority voting
-                    if not result['error_message']:
-                        results[rs] = result['result']
-                        break
-
-        for idx, output in enumerate(outputs):
-            if not output.endswith(CODE_SEPARATORS[-1]) or not self.error_recovery.majority_voting:
-                continue
-            result, _ = futures[idx].result()
-            if result['error_message']:
-                continue
-            results[idx] = result['result']
-
-        # majority voting on valid code output results
-        # if majority voting is disabled, we just take the first valid output
-        counts = Counter(res for res in results if res)
-        # all errors
-        if not counts:
-            return
-
-        most_common = counts.most_common(1)[0][0]
-        valid_idx = results.index(most_common)
-        new_output['full_prompt'].generated_solution += outputs[valid_idx]
-        new_output['error_message'] = ''
-
-        return most_common
-
-    def _send_request(self, request):
-        # temperature of 0 means greedy, but it's not always supported by the server
-        # so setting explicit greedy parameters instead
-        if request["temperature"] == 0:
-            request["temperature"] = 1.0
-            request["top_k"] = 1
-            request["top_p"] = 1.0
 
         if self.ssh_server and self.ssh_key_path:
             import sshtunnel_requests
 
-            sshtunnel_request = sshtunnel_requests.from_url(f"ssh://{self.ssh_server}:22", self.ssh_key_path)
-            outputs = sshtunnel_request.put(
-                url="http://{}:{}/generate".format(self.server_host, self.server_port),
-                data=json.dumps(request),
-                headers={"Content-Type": "application/json"},
-            ).json()
+            self.requests_lib = sshtunnel_requests.from_url(f"ssh://{self.ssh_server}:22", self.ssh_key_path)
         else:
-            outputs = requests.put(
-                url="http://{}:{}/generate".format(self.server_host, self.server_port),
-                data=json.dumps(request),
-                headers={"Content-Type": "application/json"},
-            ).json()
+            self.requests_lib = requests
 
-        return outputs
+    @abc.abstractmethod
+    def generate(
+        self,
+        prompts: list[str],
+        tokens_to_generate: int = 512,
+        temperature: float = 0.0,
+        top_p: float = 0.95,
+        top_k: int = 0,
+        repetition_penalty: float = 1.0,
+        random_seed: int = 0,
+        stop_phrases: list[str] | None = None,
+        remove_stop_phrases: bool = True,
+    ) -> list[dict]:
+        pass
 
 
 class TensorRTLLMModel(BaseModel):
-    def _single_call(
+    """Note that the current implementation supports inflight-batching so
+    to make the most use of it, you should submit a large number of prompts
+    at the same time.
+
+    A good default value is 16-32 times bigger than the model's max batch size.
+    """
+
+    def generate(
         self,
-        prompts,
-        tokens_to_generate,
-        temperature,
-        top_p,
-        top_k,
-        repetition_penalty,
-        random_seed,
-        stop_phrases: List[str],
-    ):
-        string_prompts = [str(prompt) for prompt in prompts]
+        prompts: list[str],
+        tokens_to_generate: int = 512,
+        temperature: float = 0.0,
+        top_p: float = 0.95,
+        top_k: int = 0,
+        repetition_penalty: float = 1.0,
+        random_seed: int = 0,
+        stop_phrases: list[str] | None = None,
+        remove_stop_phrases: bool = True,
+    ) -> list[dict]:
+        if stop_phrases is None:
+            stop_phrases = []
         request = {
-            "prompts": string_prompts,
             "tokens_to_generate": tokens_to_generate,
             "temperature": temperature,
             "top_k": top_k,
@@ -358,47 +129,57 @@ class TensorRTLLMModel(BaseModel):
             "repetition_penalty": repetition_penalty,
             "stop_words_list": stop_phrases,
         }
-        return self._send_request(request)
+        preprocess_request(request)
+
+        generation_ids = []
+
+        for prompt in prompts:
+            request["prompt"] = prompt
+            generation_ids.append(
+                self.requests_lib.put(
+                    url="http://{}:{}/start_generation".format(self.server_host, self.server_port),
+                    data=json.dumps(request),
+                    headers={"Content-Type": "application/json"},
+                ).json()
+            )
+
+        outputs = [None] * len(generation_ids)
+        finished_count = 0
+        while finished_count < len(generation_ids):
+            time.sleep(0.1)
+            for pos, generation_id in enumerate(generation_ids):
+                if outputs[pos] is not None:
+                    continue
+                result = self.requests_lib.put(
+                    url="http://{}:{}/get_result".format(self.server_host, self.server_port),
+                    data=json.dumps({'generation_id': generation_id}),
+                    headers={"Content-Type": "application/json"},
+                ).json()
+                if result is not None:
+                    finished_count += 1
+                    outputs[pos] = {'generation': result}
+        if remove_stop_phrases:
+            postprocess_output(outputs, stop_phrases)
+        return outputs
 
 
 class NemoModel(BaseModel):
-    def __init__(
+    def generate(
         self,
-        **kwargs,
-    ):
-        # nemo inference handles code execution directly
-        if 'handle_code_execution' not in kwargs:
-            kwargs['handle_code_execution'] = False
-        if not kwargs['handle_code_execution']:
-            unsupported_arguments = [
-                'max_code_output_characters',
-                'code_execution_timeout',
-                'max_code_executions',
-                'error_recovery',
-                'stop_on_code_error',
-            ]
-            for arg in unsupported_arguments:
-                if arg in kwargs:
-                    raise ValueError(
-                        f"`{arg}` is not supported by NemoModel if handle_code_execution=False. To use it, set handle_code_execution=True."
-                    )
-
-        super().__init__(**kwargs)
-
-    def _single_call(
-        self,
-        prompts,
-        tokens_to_generate,
-        temperature,
-        top_p,
-        top_k,
-        repetition_penalty,
-        random_seed,
-        stop_phrases: List[str],
-    ):
-        string_prompts = [str(prompt) for prompt in prompts]
+        prompts: list[str],
+        tokens_to_generate: int = 512,
+        temperature: float = 0.0,
+        top_p: float = 0.95,
+        top_k: int = 0,
+        repetition_penalty: float = 1.0,
+        random_seed: int = 0,
+        stop_phrases: list[str] | None = None,
+        remove_stop_phrases: bool = True,
+    ) -> list[dict]:
+        if stop_phrases is None:
+            stop_phrases = []
         request = {
-            "sentences": string_prompts,
+            "sentences": prompts,
             "tokens_to_generate": tokens_to_generate,
             "temperature": temperature,
             "top_k": top_k,
@@ -407,14 +188,22 @@ class NemoModel(BaseModel):
             "repetition_penalty": repetition_penalty,
             "end_strings": ["<|endoftext|>"] + stop_phrases,
         }
-        outputs = self._send_request(request)
-        outputs = outputs['sentences']
-        # always returns full prompt, so we need to remove the original prompt
-        for idx, output in enumerate(outputs):
-            outputs[idx] = output[len(string_prompts[idx]) :]
+        preprocess_request(request)
+        generations = self.requests_lib.put(
+            url="http://{}:{}/generate".format(self.server_host, self.server_port),
+            data=json.dumps(request),
+            headers={"Content-Type": "application/json"},
+        ).json()
+        # we need to remove the original prompt as nemo always returns it
+        outputs = [None] * len(generations['sentences'])
+        for idx, generation in enumerate(generations['sentences']):
+            outputs[idx] = {'generation': generation[len(prompts[idx]) :]}
+        if remove_stop_phrases:
+            postprocess_output(outputs, stop_phrases)
         return outputs
 
 
+# TODO: this is broken
 class OpenAIModel(BaseModel):
     def __init__(
         self,
@@ -431,24 +220,28 @@ class OpenAIModel(BaseModel):
         self.model = model
         self.client = OpenAI(api_key=api_key)
 
-    def _single_call(
+    def generate(
         self,
-        prompts,
-        tokens_to_generate,
-        temperature,
-        top_p,
-        repetition_penalty,
-        random_seed,
-        stop_phrases: List[str],
-        top_k=0,  # is not supported by OpenAI
-    ):
+        prompts: list[str],
+        tokens_to_generate: int = 512,
+        temperature: float = 0.0,
+        top_p: float = 0.95,
+        top_k: int = 0,
+        repetition_penalty: float = 1.0,
+        random_seed: int = 0,
+        stop_phrases: list[str] | None = None,
+        remove_stop_phrases: bool = True,
+    ) -> list[dict]:
+        if stop_phrases is None:
+            stop_phrases = []
         if top_k != 0:
             raise ValueError("`top_k` is not supported by OpenAI, please set it to default value `0`.")
 
-        responses = []
-        for prompt in prompts:
+        outputs = []
+        for input_dict in input_dicts:
             response = self._send_request(
                 prompt=prompt,
+                input_dict=input_dict,
                 tokens_to_generate=tokens_to_generate,
                 temperature=temperature,
                 top_p=top_p,
@@ -456,20 +249,24 @@ class OpenAIModel(BaseModel):
                 random_seed=random_seed,
                 stop_phrases=stop_phrases,
             )
-            responses.append(response)
-        return responses
+            outputs.append({'generation': response})
+
+        if remove_stop_phrases:
+            postprocess_output(outputs, stop_phrases)
+
+        return outputs
 
     def _send_request(
         self,
-        prompt: Prompt,
-        tokens_to_generate,
-        temperature,
-        top_p,
-        repetition_penalty,
-        random_seed,
-        stop_phrases: List[str],
-    ):
-        messages = prompt.build_chat_prompt()
+        prompt: str,
+        tokens_to_generate: int,
+        temperature: float,
+        top_p: float,
+        repetition_penalty: float,
+        random_seed: int,
+        stop_phrases: list[str],
+    ) -> str:
+        messages = prompt.build_structured(input_dict)
         response = self.client.chat.completions.create(
             model=self.model,
             temperature=temperature,
@@ -480,48 +277,19 @@ class OpenAIModel(BaseModel):
             stop=stop_phrases,
             messages=messages,
         ).choices[0]
-        content = response.message.content
-
-        # OpenAI removes stop tokens so we need to add them back
-        if (
-            response.finish_reason == "stop"
-            and content.find(CODE_SEPARATORS[0]) != -1
-            and not content[content.find(CODE_SEPARATORS[0]) :].count(CODE_SEPARATORS[-1])
-        ):
-            content += CODE_SEPARATORS[-1]
-
-        return content
+        output = response.message.content
+        # adding back stop words
+        if response.finish_reason == "stop":
+            output += response.stop_reason
+        return output
 
 
 class VLLMModel(BaseModel):
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
 
-    def __init__(
-        self,
-        host='127.0.0.1',
-        port='5000',
-        sandbox=None,
-        ssh_server=None,
-        ssh_key_path=None,
-        max_code_output_characters=1000,
-        code_execution_timeout=10.0,
-        max_code_executions=3,
-        stop_on_code_error=True,
-        handle_code_execution=True,
-        error_recovery=None,
-    ):
-        super().__init__(
-            host=host,
-            port=port,
-            sandbox=sandbox,
-            ssh_server=ssh_server,
-            ssh_key_path=ssh_key_path,
-            max_code_output_characters=max_code_output_characters,
-            code_execution_timeout=code_execution_timeout,
-            max_code_executions=max_code_executions,
-            stop_on_code_error=stop_on_code_error,
-            handle_code_execution=handle_code_execution,
-            error_recovery=error_recovery,
-        )
+        if self.ssh_server and self.ssh_key_path:
+            raise NotImplementedError("SSH tunnelling is not implemented for vLLM model.")
 
         self.server_type = "openai"
         self.oai_client = None
@@ -532,17 +300,20 @@ class VLLMModel(BaseModel):
 
         LOG.info("Model hosted by %s server: %s", self.server_type, self.model)
 
-    def _single_call(
+    def generate(
         self,
-        prompts,
-        tokens_to_generate,
-        temperature,
-        top_p,
-        top_k,
-        repetition_penalty,
-        random_seed,
-        stop_phrases: List[str],
-    ):
+        prompts: list[str],
+        tokens_to_generate: int = 512,
+        temperature: float = 0.0,
+        top_p: float = 0.95,
+        top_k: int = 0,
+        repetition_penalty: float = 1.0,
+        random_seed: int = 0,
+        stop_phrases: list[str] | None = None,
+        remove_stop_phrases: bool = True,
+    ) -> list[dict]:
+        if stop_phrases is None:
+            stop_phrases = []
         request = {
             'prompt': prompts,
             'max_tokens': tokens_to_generate,
@@ -559,22 +330,10 @@ class VLLMModel(BaseModel):
             'logit_bias': None,
             'seed': random_seed,
         }
-
-        return self._send_request(request)
-
-    def _send_request(self, request):
-        if self.ssh_server and self.ssh_key_path:
-            raise NotImplementedError("SSH tunnelling is not implemented for vLLM OpenAI server.")
-
-        # temperature of 0 means greedy, but it's not always supported by the server
-        # so setting explicit greedy parameters instead
-        if request["temperature"] == 0:
-            request["temperature"] = 1.0
-            request["top_k"] = 1
-            request["top_p"] = 1.0
-
-        outputs = self.prompt_api(**request, parse_response=True)
-
+        preprocess_request(request)
+        outputs = [{'generation': output} for output in self.prompt_api(**request, parse_response=True)]
+        if remove_stop_phrases:
+            postprocess_output(outputs, stop_phrases)
         return outputs
 
     def prompt_api(
@@ -594,15 +353,7 @@ class VLLMModel(BaseModel):
         logit_bias: dict = None,
         seed: int = None,
         parse_response: bool = True,
-    ) -> Union["openai.types.Completion", List[str]]:
-
-        if isinstance(prompt, (list, tuple)):
-            prompt = [str(prmt) for prmt in prompt]
-
-        if temperature == 0.0:
-            temperature = 1.0
-            top_k = 1
-
+    ) -> Union[list[str], "openai.types.Completion"]:
         if top_k == 0:
             top_k = 1
 
@@ -614,7 +365,6 @@ class VLLMModel(BaseModel):
                 "spaces_between_special_tokens": False,
             }
         }
-
         response = self.oai_client.completions.create(
             model=self.model,
             prompt=prompt,
@@ -639,14 +389,18 @@ class VLLMModel(BaseModel):
         return response
 
     @classmethod
-    def parse_openai_response(cls, response: "openai.types.Completion") -> List[str]:
+    def parse_openai_response(cls, response: "openai.types.Completion") -> list[str]:
         responses = []
         if not isinstance(response, list):
             response = [response]
 
         for resp in response:
             for choice in resp.choices:
-                responses.append(choice.text)
+                output = choice.text
+                # adding back stop words
+                if choice.finish_reason == "stop":
+                    output += choice.stop_reason
+                responses.append(output)
         return responses
 
     @staticmethod
@@ -670,7 +424,7 @@ class VLLMModel(BaseModel):
 models = {
     'tensorrt_llm': TensorRTLLMModel,
     'nemo': NemoModel,
-    'openai': OpenAIModel,
+    # 'openai': OpenAIModel,
     'vllm': VLLMModel,
 }
 
@@ -679,9 +433,3 @@ def get_model(server_type, **kwargs):
     """A helper function to make it easier to set server through cmd."""
     model_class = models[server_type.lower()]
     return model_class(**kwargs)
-
-
-def server_params():
-    """Returns server documentation (to include in cmd help)."""
-    prefix = f'\n        server_type: str = MISSING - Choices: {list(models.keys())}'
-    return python_doc_to_cmd_help(BaseModel, docs_prefix=prefix, arg_prefix="server.")
