@@ -24,12 +24,10 @@ import urllib.request
 from collections import defaultdict
 from pathlib import Path
 
-from normalizer import normalize_answer_string
-
 sys.path.append(str(Path(__file__).parents[1]))
 from utils import prepare_for_sft
 
-from nemo_skills.code_execution.math_grader import extract_answer
+from nemo_skills.code_execution.math_grader import extract_answer, normalize_answer_string
 
 # utils is adding main package to path already
 from nemo_skills.inference.prompt.utils import prompt_types
@@ -61,6 +59,33 @@ def _post_fix(problem_id, soln_string):
     if problem_id == "test/intermediate_algebra/78.json":
         soln_string = re.sub(r"\\(\d+)", r"\1", soln_string)
 
+    if problem_id == "train/number_theory/7115.json":
+        return "A"
+
+    if problem_id == "train/number_theory/1012.json":
+        return "E"
+
+    if problem_id == "train/prealgebra/666.json":
+        return "125"
+
+    if problem_id == "train/intermediate_algebra/172.json":
+        return "two lines"
+
+    if problem_id == "train/prealgebra/1691.json":
+        return "1.85"
+
+    if problem_id == "train/geometry/6177.json":
+        return "C"
+
+    if problem_id == "train/number_theory/7117.json":
+        return "A"
+
+    if problem_id == "train/geometry/6202.json":
+        return "D"
+
+    if problem_id == "train/precalculus/268.json":
+        return "A"
+
     return soln_string
 
 
@@ -74,7 +99,7 @@ def process_data():
     )
     parser.add_argument("--random_seed", type=int, default=42)
     parser.add_argument("--validation_size", type=int, default=1000)
-    parser.add_argument("--prompt_type", default="code_sfted", choices=prompt_types)
+    parser.add_argument("--prompt_type", default="openmathinstruct/sft", choices=prompt_types)
     args = parser.parse_args()
 
     output_folder = Path(__file__).absolute().parent
@@ -143,7 +168,10 @@ def process_data():
                 data = instances[: args.validation_size]
                 # dumping SFT-ready validation file as well right away
                 with open(output_folder / "validation-sft.jsonl", "wt", encoding="utf-8") as fout:
-                    for entry in prepare_for_sft(data, args.prompt_type, "math"):
+                    for entry in prepare_for_sft(data, args.prompt_type, "math", chat_format=False):
+                        fout.write(json.dumps(entry) + "\n")
+                with open(output_folder / "validation-sft-chat.jsonl", "wt", encoding="utf-8") as fout:
+                    for entry in prepare_for_sft(data, args.prompt_type, "math", chat_format=True):
                         fout.write(json.dumps(entry) + "\n")
             elif args.split_name == "train":
                 data = instances[args.validation_size :]
