@@ -23,19 +23,29 @@ sys.path.append(str(Path(__file__).absolute().parents[1]))
 from nemo_skills.inference.prompt.utils import Prompt, get_prompt_config
 
 
-def prepare_for_sft(data, prompt_type, dataset):
+def prepare_for_sft(data, prompt_type, dataset, chat_format=False):
     # reading prompt format from the yaml file
     prompt_config = get_prompt_config(prompt_type)
     prompt_config.context_type = "empty"
     prompt_config.few_shot_examples.num_few_shots = 0
+    prompt = Prompt(config=prompt_config)
 
     prepared_data = []
     for original_elem in data:
         elem = {}
-        elem["input"] = str(Prompt(config=prompt_config, input_dict=original_elem))
-        # note that the loss will not be meaningful,
-        # since our solution format is different, but we need to populate that field
-        elem["output"] = original_elem['reference_solution']
+        # note that the loss will not be meaningful, since we are using reference solution for output
+        # and our solution format is different, but we need to populate that field for the code to work
+        if chat_format:
+            elem['conversations'] = [
+                {'value': original_elem['question'], 'from': 'User', 'canonical_form': ''},
+                {'value': original_elem["reference_solution"], 'from': 'Assistant', 'canonical_form': ''},
+            ]
+            elem['system'] = prompt_config.system
+            elem['mask'] = 'User'
+            elem['type'] = None
+        else:
+            elem["input"] = prompt.build_string(input_dict=original_elem)
+            elem["output"] = original_elem['reference_solution']
         elem["expected_answer"] = original_elem['expected_answer']
         elem["dataset"] = dataset
         prepared_data.append(elem)
