@@ -87,9 +87,241 @@ import signal
 from math import isclose
 from typing import Union
 
-from sympy import N, simplify
-from sympy.parsing.latex import parse_latex
-from sympy.parsing.sympy_parser import parse_expr
+
+def _fix_fracs(string):
+    # replacing all extra spaces
+    while "\\frac " in string:
+        string = string.replace("\\frac ", "\\frac")
+    substrs = string.split("\\frac")
+    new_str = substrs[0]
+    if len(substrs) > 1:
+        substrs = substrs[1:]
+        for substr in substrs:
+            new_str += "\\frac"
+            if len(substr) > 0 and substr[0] == "{":
+                new_str += substr
+            else:
+                try:
+                    assert len(substr) >= 2
+                except:
+                    return string
+                a = substr[0]
+                b = substr[1]
+                if b != "{":
+                    if len(substr) > 2:
+                        post_substr = substr[2:]
+                        new_str += "{" + a + "}{" + b + "}" + post_substr
+                    else:
+                        new_str += "{" + a + "}{" + b + "}"
+                else:
+                    if len(substr) > 2:
+                        post_substr = substr[2:]
+                        new_str += "{" + a + "}" + b + post_substr
+                    else:
+                        new_str += "{" + a + "}" + b
+    string = new_str
+    return string
+
+
+def _str_is_int(x: str) -> bool:
+    try:
+        x = _strip_properly_formatted_commas(x)
+        x = float(x)
+        return abs(x - int(round(x))) <= 1e-7
+    except:
+        return False
+
+
+def _str_to_int(x: str) -> bool:
+    x = x.replace(",", "")
+    if "_" in x:
+        # Due to base
+        x = x.split("_")[0]
+    x = float(x)
+    return int(x)
+
+
+def _inject_implicit_mixed_number(step: str):
+    """
+    Automatically make a mixed number evalable
+    e.g. 7 3/4 => 7+3/4
+    """
+    p1 = re.compile("([0-9]) +([0-9])")
+    step = p1.sub("\\1+\\2", step)  ## implicit mults
+    return step
+
+
+def _strip_properly_formatted_commas(expr: str):
+    # We want to be careful because we don't want to strip tuple commas
+    p1 = re.compile("(\d)(,)(\d\d\d)($|\D)")
+    while True:
+        next_expr = p1.sub("\\1\\3\\4", expr)
+        if next_expr == expr:
+            break
+        expr = next_expr
+    return next_expr
+
+
+def _remove_right_units(expr):
+    # "\\text{ " only ever occurs (at least in the val set) when describing units
+    if "\\text" in expr:
+        try:
+            splits = re.split(r"\\text\s*{\s*", expr)
+            # print(splits)
+            assert len(splits) == 2 and splits[0] not in ("", "(")
+            return splits[0]
+        except AssertionError:
+            pass
+
+    if "\\text{" in expr:
+        return re.sub(r"\\text{([^}]+)}", r"\1", expr)
+    elif "\\mbox{" in expr:
+        splits = expr.split("\\mbox{")
+        assert len(splits) == 2
+        return splits[0]
+    else:
+        return expr
+
+
+def _process_and_or_inside_text(string):
+    string = re.sub(r"\s*\\text{\s*(or|and)\s*}\s*", ",", string)
+    string = re.sub(r",\s*,", ",", string)
+    return string
+
+
+def _remove_left_and_right(expr):
+    """Remove the right and left latex commands."""
+    expr = re.sub(r"\\left", "", expr)
+    expr = re.sub(r"\\right", "", expr)
+    return expr
+
+
+def _fix_sqrt(string):
+    _string = re.sub(r"\\sqrt(\s*\w+)", r"\\sqrt{\1}", string)
+    return _string
+
+
+def _fix_interval(expr):
+    """Fix interval expression."""
+    if "\\in " in expr:
+        return expr.split("\\in ")[1].strip()
+
+    return expr
+
+
+def _inject_implicit_mixed_fraction(step: str):
+    """
+    Automatically make a mixed number evalable
+    e.g. 7 \\frac{3}{4} => 7+3/4
+    """
+    p1 = re.compile(r"(\d+) *\\frac{(\d+)}{(\d+)}")
+
+    def replacer(match):
+        whole_part = match.group(1)
+        numerator = match.group(2)
+        denominator = match.group(3)
+
+        if whole_part:
+            return f"{whole_part} + {numerator}/{denominator}"
+        else:
+            return f"{numerator}/{denominator}"
+
+    step = p1.sub(replacer, step)
+    return step
+
+
+def normalize_answer_string(expr: str) -> str:
+    """Normalize answer expressions."""
+    if expr is None:
+        return None
+
+    # Remove enclosing `\text{}`.
+
+    expr = _remove_left_and_right(expr)
+    expr = _process_and_or_inside_text(expr)
+    expr = _remove_right_units(expr)
+    expr = _fix_interval(expr)
+    for surround_str in ["\\\\text", "\\\\mathrm", "\\\\mathcal", "\\\\textbf", "\\\\textit"]:
+        expr = expr.replace(surround_str, "")
+        pattern = f"^{surround_str}" + "\{(?P<text>.+?)\}$"
+        m = re.search(pattern, expr)
+        if m is not None:
+            expr = m.group("text")
+
+    expr = expr.replace("\!", "")
+    expr = expr.replace("\\%", "%")
+    expr = expr.replace("\\$", "$")
+    expr = expr.replace("$", "")
+    expr = expr.replace("%", "")
+    expr = expr.replace("^{\\circ}", "")
+
+    expr = expr.replace(" or ", " , ")
+    expr = expr.replace(" and ", " , ")
+
+    expr = expr.replace("million", "*10^6")
+    expr = expr.replace("billion", "*10^9")
+    expr = expr.replace("trillion", "*10^12")
+
+    for unit in [
+        "degree",
+        "cm",
+        "centimeter",
+        "meter",
+        "mile",
+        "second",
+        "minute",
+        "hour",
+        "week",
+        "month",
+        "year",
+        "foot",
+        "feet",
+        "inch",
+        "yard",
+        "p.m.",
+        "PM",
+    ]:
+        expr = re.sub(f"{unit}(es)?(s)? *(\^[0-9]+)?", "", expr)
+
+    if "day" in expr:
+        days = [
+            "Monday",
+            "Tuesday",
+            "Wednesday",
+            "Thursday",
+            "Friday",
+            "Saturday",
+            "Sunday",
+        ]
+        weekday_expressed = False
+        for day in days:
+            if day in expr:
+                weekday_expressed = True
+                break
+
+        if not weekday_expressed:
+            expr = re.sub(f"day(s)?", "", expr)
+
+    expr = re.sub(f"\^ *\\\\circ", "", expr)
+
+    if len(expr) > 0 and expr[0] == "{" and expr[-1] == "}":
+        expr = expr[1:-1]
+
+    expr = _fix_sqrt(expr)
+
+    # \frac1b or \frac12 --> \frac{1}{b} and \frac{1}{2}, etc. Even works with \frac1{72} (but not \frac{72}1). Also does a/b --> \\frac{a}{b}
+    expr = _fix_fracs(expr)
+
+    # edge case with mixed numbers and negative signs
+    expr = re.sub("- *", "-", expr)
+    expr = _inject_implicit_mixed_number(expr)
+    expr = _inject_implicit_mixed_fraction(expr)
+    expr = expr.replace(" ", "")
+
+    if _str_is_int(expr):
+        expr = str(_str_to_int(expr))
+
+    return expr
 
 
 def is_digit(s):
@@ -130,8 +362,14 @@ def math_equal(
     1. numerical equal: both can convert to float and are equal
     2. symbolic equal: both can convert to sympy expression and are equal
     """
+    from sympy.parsing.sympy_parser import parse_expr
+
     prediction = normalize(prediction)
     reference = normalize(reference)
+
+    # another round of normalization
+    prediction = normalize_answer_string(prediction)
+    reference = normalize_answer_string(reference)
 
     if isinstance(prediction, str) and len(prediction) > 1000:  # handling weird corner-cases
         prediction = prediction[:1000]
@@ -252,6 +490,10 @@ def math_equal(
 
 
 def symbolic_equal(a, b, tolerance, timeout=10.0):
+    import sympy
+    from sympy.parsing.latex import parse_latex
+    from sympy.parsing.sympy_parser import parse_expr
+
     def _parse(s):
         for f in [parse_expr, parse_latex]:
             try:
@@ -266,14 +508,14 @@ def symbolic_equal(a, b, tolerance, timeout=10.0):
 
     try:
         with time_limit(timeout):
-            if simplify(a - b) == 0:
+            if sympy.simplify(a - b) == 0:
                 return True
     except Exception:
         pass
 
     try:
         with time_limit(timeout):
-            if isclose(N(a), N(b), rel_tol=tolerance):
+            if isclose(sympy.N(a), sympy.N(b), rel_tol=tolerance):
                 return True
     except Exception:
         pass
