@@ -12,6 +12,161 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+
+# Source https://github.com/EleutherAI/lm-evaluation-harness/blob/main/lm_eval/tasks/minerva_math/utils.py#L42
+standard_four_shot = [
+    {
+        "question": "Find the domain of the expression  $\\frac{\\sqrt{x-2}}{\\sqrt{5-x}}$.}",
+        "generation": "The expressions inside each square root must be non-negative. Therefore, $x-2 \\ge 0$, so $x\\ge2$, and $5 - x \\ge 0$, so $x \\le 5$. Also, the denominator cannot be equal to zero, so $5-x>0$, which gives $x<5$. Therefore, the domain of the expression is $\\boxed{[2,5)}$.",
+    },
+    {
+        "question": "If $\\det \\mathbf{A} = 2$ and $\\det \\mathbf{B} = 12,$ then find $\\det (\\mathbf{A} \\mathbf{B}).$",
+        "generation": "We have that $\\det (\\mathbf{A} \\mathbf{B}) = (\\det \\mathbf{A})(\\det \\mathbf{B}) = (2)(12) = \\boxed{24}.$",
+    },
+    {
+        "question": "Terrell usually lifts two 20-pound weights 12 times. If he uses two 15-pound weights instead, how many times must Terrell lift them in order to lift the same total weight?",
+        "generation": "If Terrell lifts two 20-pound weights 12 times, he lifts a total of $2\\cdot 12\\cdot20=480$ pounds of weight.  If he lifts two 15-pound weights instead for $n$ times, he will lift a total of $2\\cdot15\\cdot n=30n$ pounds of weight.  Equating this to 480 pounds, we can solve for $n$:\n\\begin{align*}\n30n&=480\\\n\\Rightarrow\\qquad n&=480/30=\\boxed{16}\n\\end{align*}",
+    },
+    {
+        "problem": "If the system of equations\n\n\\begin{align*}\n6x-4y&=a,\\\n6y-9x &=b.\n\\end{align*}has a solution $(x, y)$ where $x$ and $y$ are both nonzero,\nfind $\\frac{a}{b},$ assuming $b$ is nonzero.",
+        "generation": "If we multiply the first equation by $-\\frac{3}{2}$, we obtain\n\n$$6y-9x=-\\frac{3}{2}a.$$Since we also know that $6y-9x=b$, we have\n\n$$-\\frac{3}{2}a=b\\Rightarrow\\frac{a}{b}=\\boxed{-\\frac{2}{3}}.$$",
+    },
+]
+
+
+text_detailed = [
+    {
+        "question": "A parabola with equation $y=x^2+bx+c$ passes through the points $(-1,-11)$ and $(3,17)$. What is $c$?",
+        "reference_solution": "We substitute these two points into the given equation to solve for $c$. Plugging in $(-1,-11)$, we get $-11=(-1)^2-1b+c\\Rightarrow -b+c=-12$. Plugging in $(3,17)$, we get $17=3^2+3b+c \\Rightarrow 3b+c=8$. In summary, we have the two equations \\begin{align*}\n-b+c&=-12\\\\\n3b+c&=8\n\\end{align*} Multiplying the first equation by 3, we have $-3b+3c=-36$. Adding the second equation to this last one, we have $(-3b+3c)+(3b+c)=-36+8 \\Rightarrow c=\\boxed{-7}$.",
+        "masked_reference_solution": "We substitute these two points into the given equation to solve for $c$. Plugging in $(-1,-11)$, we get $-11=(-1)^2-1b+c\\Rightarrow -b+c=-12$. Plugging in $(3,17)$, we get $17=3^2+3b+c \\Rightarrow 3b+c=8$. In summary, we have the two equations \\begin{align*}\n-b+c&=M\\\\\n3b+c&=N\n\\end{align*} Multiplying the first equation by 3, we have $-3b+3c=3M$. Adding the second equation to this last one, we have $(-3b+3c)+(3b+c)=3M + N \\Rightarrow c=\\boxed{O}$.",
+        "level": "Level 5",
+        "type": "Algebra",
+        "generation": """
+From the question we know that points $(-1, -11)$ and $(3, 17)$ lie on the parabola. This means that when we substitute $x$ and $y$ from these points into the equation $y = x^2 + bx + c$, the equation must hold true. We substitute these two points into the given equation to solve for $c$.
+
+For the point $(-1, -11)$:
+
+Substitute $x = -1$ and $ y = -11 $ into the equation:
+\\[ -11 = (-1)^2 + b(-1) + c \\Rightarrow -11 = 1 - b + c \\Rightarrow -b + c = -12 \\]
+
+For the point $(3, 17)$:
+
+Substitute $x = 3$ and $y = 17$ into the equation:
+\\[ 17 = (3)^2 + b(3) + c \\Rightarrow 17 = 9 + 3b + c \\Rightarrow 3b + c = 8 \\]
+
+In summary, we have the two equations
+\\begin{align*}
+-b + c &= -12\\\\
+3b + c &= 8
+\\end{align*}
+
+To solve for $c$ we can eliminate $b$ by multiplying the first equation by 3 and adding equations together.
+Multiplying the first equation by 3, we have $3(-b + c) = 3 (-12) \\Rightarrow -3b + 3c = -36$. Adding equations together gives us
+\\[ (-3b + 3c) + (3b + c) = -36 + 8 \\Rightarrow -3b + 3b + 3c + c = -28 \\Rightarrow 4c = -28 \\Rightarrow c = -28 : 4 \\Rightarrow c = \\boxed{-7} \\]
+""".strip(),
+    },
+    {
+        'question': 'Let $f(x)$ be an odd function.  Is $f(f(x))$ even, odd, or neither?\n\nEnter "odd", "even", or "neither".',
+        'level': 'Level 1',
+        'type': 'Intermediate Algebra',
+        'reference_solution': 'We have that\n\\[f(f(-x)) = f(-f(x)) = -f(f(x)),\\]so $f(f(x))$ is an $\\boxed{\\text{odd}}$ function.',
+        'masked_reference_solution': 'We have that\n\\[f(f(-x)) = f(-f(x)) = -f(f(x)),\\]so $f(f(x))$ is an $\\boxed{\\text{M}}$ function.',
+        'id': '954',
+        'generation': '''
+To determine whether $f(f(x))$ is even, odd, or neither, we need to use the property of $f(x)$ being an odd function.
+
+An odd function is defined as:
+\\[ f(-x) = -f(x) \\quad \\text{for all } x \\]
+
+Given that $f(x)$ is odd, let's find $f(f(-x))$ and see how it relates to $f(f(x))$.
+
+1. Substitute $-x$ into the function $f(x)$:
+\\[ f(-x) \\]
+
+2. Since $f(x)$ is odd, apply the definition of an odd function:
+\\[ f(-x) = -f(x) \\]
+
+3. Now substitute $-f(x)$ into the function $f$:
+\\[ f(f(-x)) = f(-f(x)) \\]
+
+4. Again, using the fact that $f(x)$ is odd, apply the definition:
+\\[ f(-f(x)) = -f(f(x)) \\]
+
+5. We have found that:
+\\[ f(f(-x)) = -f(f(x)) \\]
+
+This matches the definition of an odd function.
+
+So, the answer is:
+\\[ \\boxed{\\text{odd}} \\]
+'''.strip(),
+    },
+    {
+        "question": "A rectangular box $P$ is inscribed in a sphere of radius $r$. The surface area of $P$ is 384, and the sum of the lengths of its 12 edges is 112. What is $r$?",
+        "level": "Level 5",
+        "type": "Geometry",
+        "reference_solution": "Let the dimensions of $P$ be $x$, $y$, and $z$. The sum of the lengths of the edges of $P$ is $4(x+y+z)$, and the surface area of $P$ is $2xy+2yz+2xz$, so \\[\nx+y+z=28 \\quad\\text{and}\\quad 2xy+2yz+2xz=384.\n\\] Each internal diagonal of $P$ is a diameter of the sphere, so \\begin{align*}\n(2r)^2&=(x^2+y^2+z^2)\\\\\n&=(x+y+z)^2-(2xy+2xz+2yz) \\\\\n&= 28^2-384\\\\& = 400.\n\\end{align*} So $2r = 20$ and  $r=\\boxed{10}$.\n\nNote: There are infinitely many positive solutions of the system $x+y+z=28$, $2xy+2yz+2xz=384$, so there are infinitely many non-congruent boxes meeting the given conditions, but each can be inscribed in a sphere of radius 10.",
+        "masked_reference_solution": "Let the dimensions of $P$ be $x$, $y$, and $z$. The sum of the lengths of the edges of $P$ is $4(x+y+z)$, and the surface area of $P$ is $2xy+2yz+2xz$, so \\[\nx+y+z=28 \\quad\\text{and}\\quad 2xy+2yz+2xz=384.\n\\] Each internal diagonal of $P$ is a diameter of the sphere, so \\begin{align*}\n(2r)^2&=(x^2+y^2+z^2)\\\\\n&=(x+y+z)^2-(2xy+2xz+2yz) \\\\\n&= 28^2-384\\\\& = P.\n\\end{align*} So $2r = 20$ and  $r=\\boxed{N}$.\nNote: There are infinitely many positive solutions of the system $x+y+z=28$, $2xy+2yz+2xz=384$, so there are infinitely many non-congruent boxes meeting the given conditions, but each can be inscribed in a sphere of radius N.",
+        "generation": """
+Let the dimensions of the rectangular box $P$ be $x$, $y$, and $z$. We know the following:
+
+1. The sum of the lengths of the edges of $P$ is
+\\[ 4(x + y + z) = 112 \\Rightarrow x + y + z = 112 : 4 \\Rightarrow x + y + z = 28 \\]
+
+2. The surface area of $P$ is
+\\[ 2xy + 2yz + 2xz = 384 \\Rightarrow xy + yz + xz = 384 : 2 \\Rightarrow xy + yz + xz = 192 \\]
+
+Since the box is inscribed in the sphere, the diagonal of the box is the diameter of the sphere. The length of the diagonal is $\\sqrt{x^2 + y^2 + z^2}$
+
+The diameter of the sphere is $2r$, so:
+\\[ 2r = \\sqrt{x^2 + y^2 + z^2} \\Rightarrow (2r)^2 = x^2 + y^2 + z^2 = (x + y + z)^2 - (2xy + 2yz + 2xz) \\]
+
+Substitute the known values:
+\\[ 4r^2 = 28^2 - 384 = 784 - 384 = 400 \\Rightarrow r^2 = 100 \\Rightarrow r = \\boxed{10} \\]
+""".strip(),
+    },
+    {
+        "question": "Let $\\mathbf{a} = \\begin{pmatrix} 2 \\\\ 1 \\\\ 5 \\end{pmatrix}.$  Find the vector $\\mathbf{b}$ such that $\\mathbf{a} \\cdot \\mathbf{b} = 11$ and\n\\[\\mathbf{a} \\times \\mathbf{b} = \\begin{pmatrix} -13 \\\\ -9 \\\\ 7 \\end{pmatrix}.\\]",
+        "level": "Level 3",
+        "type": "Precalculus",
+        "reference_solution": "Let $\\mathbf{b} = \\begin{pmatrix} x \\\\ y \\\\ z \\end{pmatrix}.$  Then the equation $\\mathbf{a} \\cdot \\mathbf{b} = 11$ gives us $2x + y + 5z = 11.$  Also,\n\\[\\mathbf{a} \\times \\mathbf{b} = \\begin{pmatrix} 2 \\\\ 1 \\\\ 5 \\end{pmatrix} \\times \\begin{pmatrix} x \\\\ y \\\\ z \\end{pmatrix} = \\begin{pmatrix} -5y + z \\\\ 5x - 2z \\\\ -x + 2y \\end{pmatrix}.\\]Comparing entries, we obtain\n\\begin{align*}\n-5y + z &= -13, \\\\\n5x - 2z &= -9, \\\\\n-x + 2y &= 7.\n\\end{align*}Solving this system, along with the equation $2x + y + z = 5z = 11,$ we find $x = -1,$ $y = 3,$ and $z = 2.$  Hence, $\\mathbf{b} = \\boxed{\\begin{pmatrix} -1 \\\\ 3 \\\\ 2 \\end{pmatrix}}.$",
+        "masked_reference_solution": "Let $\\mathbf{b} = \\begin{pmatrix} x \\\\ y \\\\ z \\end{pmatrix}.$  Then the equation $\\mathbf{a} \\cdot \\mathbf{b} = 11$ gives us $M_1 * x + M_2 * y + M_3 * z = 11.$  Also,\n\\[\\mathbf{a} \\times \\mathbf{b} = \\begin{pmatrix} 2 \\\\ 1 \\\\ 5 \\end{pmatrix} \\times \\begin{pmatrix} x \\\\ y \\\\ z \\end{pmatrix} = \\begin{pmatrix} N_01 * y + N_02 * z \\\\ N_11 * x + N12 * z \\\\ N_21 * x + N_22 * y \\end{pmatrix}.\\]Comparing entries, we obtain\n\\begin{align*}\nN_01 * y + N_02 * z &= -13, \\\\\n N_11 * x + N12 * z &= -9, \\\\\n N_21 * x + N_22 * y &= 7.\n\\end{align*}Solving this system, along with the equation $M_1 * x + M_2 * y + M_3 * z = 11$ we find $x = O_0,$ $y = O_1,$ and $z = O_2.$  Hence, $\\mathbf{b} = \\boxed{\\begin{pmatrix} O_0 \\\\ O_1 \\\\ O_2 \\end{pmatrix}}.$",
+        "generation": """
+Let $\\mathbf{b} = \\begin{pmatrix} x \\\\ y \\\\ z \\end{pmatrix}$.
+
+First, use the dot product condition:
+\\[ \\mathbf{a} \\cdot \\mathbf{b} = 11 \\Rightarrow 2x + y + 5z = 11 \\]
+
+Next, use the cross product condition:
+\\[ \\mathbf{a} \\times \\mathbf{b} = \\begin{pmatrix} 2 \\\\ 1 \\\\ 5 \\end{pmatrix} \\times \\begin{pmatrix} x \\\\ y \\\\ z \\end{pmatrix} = \\begin{pmatrix} -5y + z \\\\ 5x - 2z \\\\ -x + 2y \\end{pmatrix} = \\begin{pmatrix} -13 \\\\ -9 \\\\ 7 \\end{pmatrix} \\]
+
+This gives us the system of equations:
+   \\begin{align*}
+   2x + y + 5z = 11 \\quad &(1) \\\\
+   -5y + z = -13 \\quad &(2) \\\\
+   5x - 2z = -9 \\quad &(3) \\\\
+   -x + 2y = 7 \\quad &(4)
+   \\end{align*}
+
+Solve for $x$, $y$, and $z$ step-by-step:
+
+From (2), $z = 5y - 13$.
+From (4), $x = 2y - 7$.
+
+Substitute $z = 5y - 13$ into (1):
+\\[ 2(2y - 7) + y + 5(5y - 13) = 11 \\Rightarrow 4y - 14 + y + 25y - 65 = 11 \\Rightarrow 30y - 79 = 11 \\Rightarrow 30y = 90 \\Rightarrow y = 3 \\]
+
+Now find $x$ and $z$:
+\\[ x = 2y - 7 = 2(3) - 7 = -1 \\]
+
+\\[ z = 5y - 13 = 5(3) - 13 = 2 \\]
+
+Thus, the vector $\\mathbf{b}$ is:
+\\[ \\mathbf{b} = \\boxed{\\begin{pmatrix} -1 \\\\ 3 \\\\ 2 \\end{pmatrix}} \\]
+""".strip(),
+    },
+]
+
 text_with_code = [
     {
         "question": "A parabola with equation $y=x^2+bx+c$ passes through the points $(-1,-11)$ and $(3,17)$. What is $c$?",
@@ -19,7 +174,7 @@ text_with_code = [
         "masked_reference_solution": "We substitute these two points into the given equation to solve for $c$. Plugging in $(-1,-11)$, we get $-11=(-1)^2-1b+c\\Rightarrow -b+c=-12$. Plugging in $(3,17)$, we get $17=3^2+3b+c \\Rightarrow 3b+c=8$. In summary, we have the two equations \\begin{align*}\n-b+c&=M\\\\\n3b+c&=N\n\\end{align*} Multiplying the first equation by 3, we have $-3b+3c=3M$. Adding the second equation to this last one, we have $(-3b+3c)+(3b+c)=3M + N \\Rightarrow c=\\boxed{O}$.",
         "level": "Level 5",
         "type": "Algebra",
-        "generated_solution": """
+        "generation": """
 Let's write down an equation for the parabola and solve for $c$ using sympy.
 {start_code}
 import sympy as sp
@@ -53,7 +208,7 @@ So c is \\boxed{-7}
         'reference_solution': 'We have that\n\\[f(f(-x)) = f(-f(x)) = -f(f(x)),\\]so $f(f(x))$ is an $\\boxed{\\text{odd}}$ function.',
         'masked_reference_solution': 'We have that\n\\[f(f(-x)) = f(-f(x)) = -f(f(x)),\\]so $f(f(x))$ is an $\\boxed{\\text{M}}$ function.',
         'id': '954',
-        'generated_solution': '''
+        'generation': '''
 The definition of an odd function is that $f(-x) = -f(x)$.
 Applying this to $f(f(-x))$ we get $f(f(-x)) = f(-f(x)) = -f(f(x))$.
 Thus, $f(f(x))$ is an \\boxed{odd} function.
@@ -65,7 +220,7 @@ Thus, $f(f(x))$ is an \\boxed{odd} function.
         "type": "Prealgebra",
         "reference_solution": "At the 2007 Math Olympics, Canada won $17$ of $100$ possible medals, or $0.17$ of the possible medals. We convert each of the possible answers to a decimal and see which is closest to $0.17:$ \\[\\frac{1}{4}=0.25 \\quad\n\\frac{1}{5}=0.2 \\quad\n\\frac{1}{6}=0.166666... \\quad\n\\frac{1}{7}=0.142857... \\quad\n\\frac{1}{8}=0.125 \\]The choice that is closest to $0.17$ is $\\boxed{\\frac{1}{6}}.$",
         "masked_reference_solution": "At the 2007 Math Olympics, Canada won $17$ of $100$ possible medals, or $0.17$ of the possible medals. We convert each of the possible answers to a decimal and see which is closest to $0.17:$ \\[\\frac{1}{4}=M \\quad\n\\frac{1}{5}=N \\quad\n\\frac{1}{6}=O \\quad\n\\frac{1}{7}=P \\quad\n\\frac{1}{8}=Q \\]The choice that is closest to $0.17$ is $\\boxed{R}$.",
-        "generated_solution": """
+        "generation": """
 Let's use sympy to print out the differences between the fraction of medals that Canada won and each of the options.
 {start_code}
 from sympy import Rational, Abs
@@ -104,7 +259,7 @@ So the answer is \\boxed{1/6}.
         "type": "Geometry",
         "reference_solution": "Let the dimensions of $P$ be $x$, $y$, and $z$. The sum of the lengths of the edges of $P$ is $4(x+y+z)$, and the surface area of $P$ is $2xy+2yz+2xz$, so \\[\nx+y+z=28 \\quad\\text{and}\\quad 2xy+2yz+2xz=384.\n\\] Each internal diagonal of $P$ is a diameter of the sphere, so \\begin{align*}\n(2r)^2&=(x^2+y^2+z^2)\\\\\n&=(x+y+z)^2-(2xy+2xz+2yz) \\\\\n&= 28^2-384\\\\& = 400.\n\\end{align*} So $2r = 20$ and  $r=\\boxed{10}$.\n\nNote: There are infinitely many positive solutions of the system $x+y+z=28$, $2xy+2yz+2xz=384$, so there are infinitely many non-congruent boxes meeting the given conditions, but each can be inscribed in a sphere of radius 10.",
         "masked_reference_solution": "Let the dimensions of $P$ be $x$, $y$, and $z$. The sum of the lengths of the edges of $P$ is $4(x+y+z)$, and the surface area of $P$ is $2xy+2yz+2xz$, so \\[\nx+y+z=28 \\quad\\text{and}\\quad 2xy+2yz+2xz=384.\n\\] Each internal diagonal of $P$ is a diameter of the sphere, so \\begin{align*}\n(2r)^2&=(x^2+y^2+z^2)\\\\\n&=(x+y+z)^2-(2xy+2xz+2yz) \\\\\n&= 28^2-384\\\\& = P.\n\\end{align*} So $2r = 20$ and  $r=\\boxed{N}$.\nNote: There are infinitely many positive solutions of the system $x+y+z=28$, $2xy+2yz+2xz=384$, so there are infinitely many non-congruent boxes meeting the given conditions, but each can be inscribed in a sphere of radius N.",
-        "generated_solution": """
+        "generation": """
 Let the dimensions of $P$ be $x$, $y$, and $z$.
 The sum of the box's edges is $4(x + y + z)$ and the surface area is $2xy + 2yz + 2xz$.
 The diameter of the sphere is $2r$ and it's equal to the diagonal of the box.
@@ -140,7 +295,7 @@ Since the radius of the sphere is positive, we get $r = \\boxed{10}$.
         "type": "Precalculus",
         "reference_solution": "Let $\\omega = e^{\\pi i/6}.$  Then assuming the bee starts at the origin, $P_{2015}$ is at the point\n\\[z = 1 + 2 \\omega + 3 \\omega^2 + 4 \\omega^3 + \\dots + 2015 \\omega^{2014}.\\]Then\n\\[\\omega z = \\omega + 2 \\omega^2 + 3 \\omega^3 + 4 \\omega^4 + \\dots + 2015 \\omega^{2015}.\\]Subtracting these equations, we get\n\\begin{align*}\n(\\omega - 1) z &= 2015 \\omega^{2015} - \\omega^{2014} - \\omega^{2013} - \\dots - \\omega - 1 \\\\\n&= 2015 \\omega^{2015} - \\frac{\\omega^{2015} - 1}{\\omega - 1}.\n\\end{align*}Since $\\omega^6 = 1, \\ $ $\\omega^{2015} = (\\omega^6)^{335} \\cdot \\omega^5 = \\omega^5.$  Hence,\n\\begin{align*}\n(\\omega - 1) z &= 2015 \\omega^5 - \\frac{\\omega^5 - 1}{\\omega - 1} \\\\\n&= 2015 \\omega^5 - \\omega^4 - \\omega^3 - \\omega^2 - \\omega - 1.\n\\end{align*}And since $\\omega^3 = -1,$ this reduces to\n\\begin{align*}\n(\\omega - 1) z &= -2015 \\omega^2 + \\omega + 1 - \\omega^2 - \\omega - 1 \\\\\n&= -2015 \\omega^2 - \\omega^2 = -2016 \\omega^2,\n\\end{align*}so\n\\[z = -\\frac{2016 \\omega^2}{\\omega - 1}.\\]Hence,\n\\[|z| = \\left|  -\\frac{2016 \\omega^2}{\\omega - 1} \\right| = \\frac{2016}{|\\omega - 1|}.\\]If we plot 0, 1, and $\\omega$ in the complex plane, we obtain an isosceles triangle.\n\n[asy]\nunitsize(4 cm);\n\npair M, O, P, Q;\n\nO = (0,0);\nP = (1,0);\nQ = dir(30);\nM = (P + Q)/2;\n\ndraw(O--P--Q--cycle);\ndraw(O--M);\n\nlabel(\"$0$\", O, SW);\nlabel(\"$1$\", P, SE);\nlabel(\"$\\omega$\", Q, NE);\nlabel(\"$1$\", (O + P)/2, S, red);\nlabel(\"$1$\", (O + Q)/2, NW, red);\n[/asy]\n\nThus, the distance between 1 and $\\omega$ is $|\\omega - 1| = 2 \\sin \\frac{\\pi}{12} = \\frac{\\sqrt{6} - \\sqrt{2}}{2},$ so\n\\[|z| = \\frac{2016}{\\frac{\\sqrt{6} - \\sqrt{2}}{2}} =\\frac{4032}{\\sqrt{6} - \\sqrt{2}} = \\frac{4032 (\\sqrt{6} + \\sqrt{2})}{4} = \\boxed{1008 \\sqrt{6} + 1008 \\sqrt{2}}.\\]",
         "masked_reference_solution": "Let $\\omega = e^{\\pi i/6}.$  Then assuming the bee starts at the origin, $P_{2015}$ is at the point\n\\[z = 1 + 2 \\omega + 3 \\omega^2 + 4 \\omega^3 + \\dots + 2015 \\omega^{2014}.\\]Then\n\\[\\omega z = \\omega + 2 \\omega^2 + 3 \\omega^3 + 4 \\omega^4 + \\dots + 2015 \\omega^{2015}.\\]Subtracting these equations, we get\n\\begin{align*}\n(\\omega - 1) z &= 2015 \\omega^{2015} - \\omega^{2014} - \\omega^{2013} - \\dots - \\omega - 1 \\\\\n&= 2015 \\omega^{2015} - \\frac{\\omega^{2015} - 1}{\\omega - 1}.\n\\end{align*}Since $\\omega^6 = 1, \\ $ $\\omega^{2015} = (\\omega^6)^{335} \\cdot \\omega^5 = \\omega^5.$  Hence,\n\\begin{align*}\n(\\omega - 1) z &= 2015 \\omega^5 - \\frac{\\omega^5 - 1}{\\omega - 1} \\\\\n&= 2015 \\omega^5 - \\omega^4 - \\omega^3 - \\omega^2 - \\omega - 1.\n\\end{align*}And since $\\omega^3 = -1,$ this reduces to\n\\begin{align*}\n(\\omega - 1) z &= -2015 \\omega^2 + \\omega + 1 - \\omega^2 - \\omega - 1 \\\\\n&= -2015 \\omega^2 - \\omega^2 = -2016 \\omega^2,\n\\end{align*}so\n\\[z = -\\frac{2016 \\omega^2}{\\omega - 1}.\\]Hence,\n\\[|z| = \\left|  -\\frac{2016 \\omega^2}{\\omega - 1} \\right| = \\frac{2016}{|\\omega - 1|}.\\]If we plot 0, 1, and $\\omega$ in the complex plane, we obtain an isosceles triangle.\n\n[asy]\nunitsize(4 cm);\n\npair M, O, P, Q;\n\nO = (0,0);\nP = (1,0);\nQ = dir(30);\nM = (P + Q)/2;\n\ndraw(O--P--Q--cycle);\ndraw(O--M);\n\nlabel(\"$0$\", O, SW);\nlabel(\"$1$\", P, SE);\nlabel(\"$\\omega$\", Q, NE);\nlabel(\"$1$\", (O + P)/2, S, red);\nlabel(\"$1$\", (O + Q)/2, NW, red);\n[/asy]\n\nThus, the distance between 1 and $\\omega$ is $|\\omega - 1| = 2 \\sin \\frac{\\pi}{12} = \\frac{\\sqrt{6} - \\sqrt{2}}{2},$ so\n\\[|z| = \\frac{2016}{\\frac{\\sqrt{6} - \\sqrt{2}}{2}} =\\frac{4032}{\\sqrt{6} - \\sqrt{2}} = \\frac{4032 (\\sqrt{6} + \\sqrt{2})}{4} = \\boxed{M}.\\]",
-        "generated_solution": """
+        "generation": """
 We can represent the rotation action via $\\omega = e^{\\pi i/6}$.
 Let's assume the bee starts at the origin, so $P_{2015}$ is at the point \\[z = 1 + 2 \\omega + 3 \\omega^2 + 4 \\omega^3 + \\dots + 2015 \\omega^{2014}.\\]
 This is an arithmetic-geometric series which we can solve by simplifying the expression.
@@ -178,7 +333,7 @@ examples_algebra = [
         "type": "Algebra",
         "reference_solution": "Let $j$ be John's age and $d$ be his dad's age. We are trying to find the value of $j$. We can create a system of two equations to represent the given information. They are\n\n\\begin{align*}\nj &= d - 24 \\\\\nj + d &= 68 \\\\\n\\end{align*}We want to find $j$, so we need to eliminate $d$ from the equations above. Rewriting the first equation we get $d = j+24$. Substituting this into the second equation to eliminate $d$, we have $j+(j+24)=68$, or $j=22$. Thus, John is $\\boxed{22}$ years old.",
         "masked_reference_solution": "Let $j$ be John's age and $d$ be his dad's age. We are trying to find the value of $j$. We can create a system of two equations to represent the given information. They are\n\n\\begin{align*}\nj &= d - 24 \\\\\nj + d &= 68 \\\\\n\\end{align*}We want to find $j$, so we need to eliminate $d$ from the equations above. Rewriting the first equation we get $d = j+24$. Substituting this into the second equation to eliminate $d$, we have $j+(j+24)=68$, or $j=M$. Thus, John is $\\boxed{M}$ years old.",
-        "generated_solution": """
+        "generation": """
 To compute John's age, we can use Python's sympy library.
 {start_code}
 import sympy as sp
@@ -207,7 +362,7 @@ So John is \\boxed{22} years old.
         "type": "Algebra",
         "reference_solution": "The given expression can be rewritten as $2x+8x^2+9-4+2x+8x^2$. Combining like terms, this last expression is equal to $(2x+2x)+(8x^2+8x^2)+(9-4)=\\boxed{16x^2+4x+5}$.",
         "masked_reference_solution": "The given expression can be rewritten as $2x+8x^2+9-4+2x+8x^2$. Combining like terms, this last expression is equal to $(2x+2x)+(8x^2+8x^2)+(9-4)=\\boxed{Mx^2+Nx+O}$.",
-        "generated_solution": """
+        "generation": """
 To simplify the expression, we can use Python's sympy library.
 {start_code}
 from sympy import symbols, simplify
@@ -230,7 +385,7 @@ So the simplified expression is $\\boxed{16x^2+4x+5}$
         "type": "Algebra",
         "reference_solution": "In this particular case, the fraction will be undefined only if its denominator is equal to zero. Because of this, we can ignore the numerator. We start by setting the binomial in the denominator equal to 0: \\begin{align*} 8x^2-65x+8=0\n\\\\\\Rightarrow\\qquad (8x-1)(x-8)=0\n\\end{align*} We find that the two possible values for $x$ are $\\frac{1}{8}$ and $8$. Since the question asks for the largest value, the final solution is $\\boxed{8}$.",
         "masked_reference_solution": "In this particular case, the fraction will be undefined only if its denominator is equal to zero. Because of this, we can ignore the numerator. We start by setting the binomial in the denominator equal to 0: \\begin{align*} 8x^2-65x+8=0\n\\\\\\Rightarrow\\qquad (8x-1)(x-8)=0\n\\end{align*} We find that the two possible values for $x$ are M and N. Since the question asks for the largest value, the final solution is $\\boxed{N}$.",
-        "generated_solution": """
+        "generation": """
 To solve this problem, we can use Python's sympy library.
 {start_code}
 from sympy import Symbol, simplify, solve
@@ -261,7 +416,7 @@ So the largest value of $x$ for which the expression is undefined is \\boxed{8}
         "type": "Algebra",
         "reference_solution": "We first evaluate $f(5) = 5 -t(5) = 5-\\sqrt{5\\cdot3+1}=1$. Thus $t(f(5))=t(1)=\\sqrt{3\\cdot1 + 1}=\\boxed{2}$.",
         "masked_reference_solution": "We first evaluate $f(5) = 5 -t(5) = 5-\\sqrt{5\\cdot3+1}=M$. Thus $t(f(5))=t(M)=\\sqrt{3\\cdot M + 1}=\\boxed{N}$.",
-        "generated_solution": """
+        "generation": """
 To compute t(f(5)), we first need to compute f(5), which in turn requires computing t(5) first.
 To compute this function composition, we use Python's sympy library.
 {start_code}
@@ -291,7 +446,7 @@ So $t(f(5))$ is \\boxed{2}
         "type": "Algebra",
         "reference_solution": "We need some basic facts from number theory: $a^0 = 1$ for any $a,$ $1^b = 1$ for any $b,$ and $(-1)^c = 1$ if $c$ is an even integer. Unless the base is a complex number (which is excluded since we are looking for integer solutions), there are no other ways to get an RHS of $1.$ Thus, either the exponent is zero $($giving the equation $25 - x^2 = 0),$ the base is $1$ $($giving $x -2 = 1),$ or the base is $-1$ and the exponent is even $($giving the simultaneous equations $x - 2 = -1$ and $25 - x^2 = 2n$ for some integer $n).$ Solving the first equation gives $x = \\pm 5,$ and solving the second gives $x = 3.$ The third equation implies that $x = 1,$ in which case $25 - x^2 = 24$ is indeed even, so $x = 1$ is a valid solution. In all, there are $\\boxed{4}$ integer solutions.",
         "masked_reference_solution": "We need some basic facts from number theory: $a^0 = 1$ for any $a,$ $1^b = 1$ for any $b,$ and $(-1)^c = 1$ if $c$ is an even integer. Unless the base is a complex number (which is excluded since we are looking for integer solutions), there are no other ways to get an RHS of $1.$ Thus, either the exponent is zero $($giving the equation $25 - x^2 = 0),$ the base is $1$ $($giving $x -2 = 1),$ or the base is $-1$ and the exponent is even $($giving the simultaneous equations $x - 2 = -1$ and $25 - x^2 = 2n$ for some integer $n).$ Solving the first equation gives $x = \\pm M,$ and solving the second gives $x = N.$ The third equation implies that $x = O,$ in which case $25 - x^2 = P$ is indeed even, so $x = O$ is a valid solution. In all, there are $\\boxed{P}$ integer solutions.",
-        "generated_solution": """
+        "generation": """
 Let's first solve the equation using sympy.
 {start_code}
 from sympy import symbols, Eq, solve
@@ -323,7 +478,7 @@ examples_probability = [
         "type": "Counting & Probability",
         "reference_solution": "The spinner is guaranteed to land on exactly one of the three regions, so we know that the sum of the probabilities of it landing in each region will be 1. If we let the probability of it landing in region $C$ be $x$, we then have the equation $1 = \\frac{5}{12}+\\frac{1}{3}+x$, from which we have $x=\\boxed{\\frac{1}{4}}$.",
         "masked_reference_solution": "The spinner is guaranteed to land on exactly one of the three regions, so we know that the sum of the probabilities of it landing in each region will be 1. If we let the probability of it landing in region $C$ be $x$, we then have the equation $1 = \\frac{5}{12}+\\frac{1}{3}+x$, from which we have $x=\\boxed{\\frac{M}{N}}$.",
-        "generated_solution": """
+        "generation": """
 Since landing on the three parts are mutually exclusive events.
 To compute the probability of landing on C, we can simply subtract the probability of landing on A and B.
 {start_code}
@@ -352,7 +507,7 @@ So probability of the event is $\\boxed{\\frac{1}{4}}$.
         "type": "Counting & Probability",
         "reference_solution": "There are $\\binom{11}{2} = 55$ combinations of two balls that can be drawn.  There are $\\binom{5}{2} = 10$ combinations of two white balls that can be drawn.  So the probability that two balls pulled out are both white is $\\dfrac{10}{55} = \\boxed{\\dfrac{2}{11}}$.",
         "masked_reference_solution": "There are $\\binom{M}{2} = N$ combinations of two balls that can be drawn.  There are $\\binom{5}{2} = O$ combinations of two white balls that can be drawn.  So the probability that two balls pulled out are both white is $\\dfrac{O}{N} = \\boxed{\\dfrac{Q}{R}}$.",
-        "generated_solution": """
+        "generation": """
 To solve the problem, we can use Python's sympy library.
 {start_code}
 from sympy import Rational, binomial
@@ -389,7 +544,7 @@ So probability of the event is $\\boxed{\\frac{2}{11}}$.
         "type": "Counting & Probability",
         "reference_solution": "There are $\\binom{5}{2} = 10$ different pairs of marbles can be drawn, and the expected value of the sum is the average of the sums of each pair.  This is  \\begin{align*}\n\\frac{1}{10}((1+2)+(1+3)+(1+4)+(1+5)+(2+3)&\\\\\n+(2+4)+(2+5)+(3+4)+(3+5)+(4+5))&=\\frac{60}{10} = \\boxed{6}. \\end{align*}",
         "masked_reference_solution": "There are $\\binom{5}{2} = M$ different pairs of marbles can be drawn, and the expected value of the sum is the average of the sums of each pair.  This is  \\begin{align*}\n\\frac{1}{M}((1+2)+(1+3)+(1+4)+(1+5)+(2+3)&\\\\\n+(2+4)+(2+5)+(3+4)+(3+5)+(4+5))&=\\frac{N}{M} = \\boxed{O}. \\end{align*}",
-        "generated_solution": """
+        "generation": """
 To calculate the expected value of the sum of two marbles, we need to iterate over all possible combinations.
 To facilitate this computation, we will use Python's numerical libraries.
 {start_code}
@@ -424,7 +579,7 @@ So the expected value is \\boxed{6}
         "type": "Counting & Probability",
         "reference_solution": "There is 1 way to make the first letter C and 3 ways to make one of the other letters B. We now have 4 ways to pick the letter for the first remaining spot and 3 ways to pick the letter for the last remaining spot. This is a total of $1\\cdot3\\cdot4\\cdot3=\\boxed{36}$ ways to arrange the letters.",
         "masked_reference_solution": "There is 1 way to make the first letter C and 3 ways to make one of the other letters B. We now have M ways to pick the letter for the first remaining spot and N ways to pick the letter for the last remaining spot. This is a total of $1\\cdot3\\cdotM\\cdotN=\\boxed{O}$ ways to arrange the letters.",
-        "generated_solution": """
+        "generation": """
 Given that the first letter is C, to construct a four-letter word with unique characters, there are three remaining choices.
 Out of these three letters, one of the letter has to be B.
 To calculate the number of possible arrangements, we will use Python's sympy library.
@@ -456,7 +611,7 @@ So the total arrangements are \\boxed{36}
         "type": "Counting & Probability",
         "reference_solution": "She can do this if and only if at least one of the dice lands on a 1. The probability neither of the dice is a 1 is $\\left(\\frac{5}{6}\\right) \\left(\\frac{5}{6}\\right) = \\frac{25}{36}$. So the probability at least one die is a 1 is $1-\\frac{25}{36} = \\boxed{\\frac{11}{36}}$.",
         "masked_reference_solution": "She can do this if and only if at least one of the dice lands on a 1. The probability neither of the dice is a 1 is $\\left(\\frac{5}{6}\\right) \\left(\\frac{5}{6}\\right) = \\frac{M}{N}$. So the probability at least one die is a 1 is $1-\\frac{M}{N} = \\boxed{\\frac{O}{P}}$.",
-        "generated_solution": """
+        "generation": """
 Melinda can never construct 10 and 20 because there's no 0 on a dice.
 To construct numbers between 10 and 20, she needs a 1 on at least one of the dices.
 To calculate these odds we will use Python's sympy library.
@@ -488,7 +643,7 @@ examples_intermediate_algebra = [
         "reference_solution": "After the subtractions are performed, each fraction in the pattern has a numerator that is one less than its denominator. The product then reduces quite nicely, leaving just the frst numerator and the last denominator, as follows: $\\frac{1}{2}\\times\\frac{2}{3}\\times\\frac{3}{4}\\times \\cdots\\times\\frac{49}{50} = \\boxed{\\frac{1}{50}}$.",
         "masked_reference_solution": "After the subtractions are performed, each fraction in the pattern has a numerator that is one less than its denominator. The product then reduces quite nicely, leaving just the frst numerator and the last denominator, as follows: $\\frac{1}{2}\\times\\frac{2}{3}\\times\\frac{3}{4}\\times \\cdots\\times\\frac{49}{50} = \\boxed{\\frac{M}{N}}$.",
         "id": "train/intermediate_algebra/1084.json",
-        "generated_solution": """
+        "generation": """
 To solve this problem, we can use Python's sympy library.
 {start_code}
 from sympy import Rational
@@ -512,7 +667,7 @@ So the answer is $\\boxed{\\frac{1}{50}}$.
         "reference_solution": "Let $p(x) = x^2-4x+3$. A number $c$ is not in the domain of $g$ if and only if $p(c) = 0$. Hence we have,\n$$c^2-4c+3=0.$$Factoring gives us\n$$(c-3)(c-1) = 0.$$Solving for $c$ gives us $1$ and $3$. Hence the domain of $g$ is $\\boxed{(-\\infty, 1) \\cup (1, 3) \\cup (3, \\infty)} $.",
         "masked_reference_solution": "Let $p(x) = x^2-4x+3$. A number $c$ is not in the domain of $g$ if and only if $p(c) = 0$. Hence we have,\n$$c^2-4c+3=0.$$Factoring gives us\n$$(c-M)(c-N) = 0.$$Solving for $c$ gives us M and N. Hence the domain of $g$ is $\\boxed{(-\\infty, M) \\cup (M, N) \\cup (N, \\infty)} $.",
         "id": "train/intermediate_algebra/2016.json",
-        "generated_solution": """
+        "generation": """
 To find the domain of the rational function
 $g(x) = \frac{x^3 - 2x^2 + 4x + 3}{x^2 - 4x + 3}$,
 we need to identify the values of $x$ for which the denominator is not equal to zero,
@@ -542,7 +697,7 @@ So the domain is: $\\boxed{ (-\infty, 1) \cup (1, 3) \cup (3, \infty) }$.
         "reference_solution": "Since $2x+7$ is a factor, we should get a remainder of $0$ when we divide $6x^3+19x^2+cx+35$.\n\\[\n\\begin{array}{c|cccc}\n\\multicolumn{2}{r}{3x^2} & -x&+5  \\\\\n\\cline{2-5}\n2x+7 & 6x^3&+19x^2&+cx&+35 \\\\\n\\multicolumn{2}{r}{-6x^3} & -21x^2  \\\\ \n\\cline{2-3}\n\\multicolumn{2}{r}{0} & -2x^2 & +cx  \\\\\n\\multicolumn{2}{r}{} & +2x^2 & +7x \\\\ \n\\cline{3-4}\n\\multicolumn{2}{r}{} & 0 & (c+7)x & + 35 \\\\ \n\\multicolumn{2}{r}{} & & -10x & -35 \\\\ \n\\cline{4-5}\n\\multicolumn{2}{r}{} & & (c+7-10)x & 0 \\\\ \n\\end{array}\n\\]The remainder is $0$ if $c+7-10=0$, so $c=\\boxed{3}$.",
         "masked_reference_solution": "Since $2x+7$ is a factor, we should get a remainder of $0$ when we divide $6x^3+19x^2+cx+35$.\n\\[\n\\begin{array}{c|cccc}\n\\multicolumn{2}{r}{3x^2} & -x&+5  \\\\\n\\cline{2-5}\n2x+7 & 6x^3&+19x^2&+cx&+35 \\\\\n\\multicolumn{2}{r}{-6x^3} & -21x^2  \\\\ \n\\cline{2-3}\n\\multicolumn{2}{r}{0} & -2x^2 & +cx  \\\\\n\\multicolumn{2}{r}{} & +2x^2 & +7x \\\\ \n\\cline{3-4}\n\\multicolumn{2}{r}{} & 0 & (c+7)x & + 35 \\\\ \n\\multicolumn{2}{r}{} & & -10x & -35 \\\\ \n\\cline{4-5}\n\\multicolumn{2}{r}{} & & (c+7-10)x & 0 \\\\ \n\\end{array}\n\\]The remainder is $0$ if $c+7-10=0$, so $c=\\boxed{M}$.",
         "id": "train/intermediate_algebra/1912.json",
-        "generated_solution": """
+        "generation": """
 We can solve this problem using sympy library.
 {start_code}
 from sympy import symbols, div, solve
@@ -575,7 +730,7 @@ So the only possible value of c is $\\boxed{3}$.
         "reference_solution": "We can write\n\\[\\frac{x^2}{x - 8} = \\frac{x^2 - 64 + 64}{x - 8} = \\frac{(x - 8)(x + 8) + 64}{x - 8} = x + 8 + \\frac{64}{x - 8} = x - 8 + \\frac{64}{x - 8} + 16.\\]By AM-GM,\n\\[x - 8 + \\frac{64}{x - 8} \\ge 2 \\sqrt{(x - 8) \\cdot \\frac{64}{x - 8}} = 16,\\]so\n\\[\\frac{x^2}{x - 8} \\ge 32.\\]Equality occurs when $x = 16,$ so the minimum value is $\\boxed{32}.$",
         "masked_reference_solution": "We can write\n\\[\\frac{x^2}{x - 8} = \\frac{x^2 - 64 + 64}{x - 8} = \\frac{(x - 8)(x + 8) + 64}{x - 8} = x + 8 + \\frac{64}{x - 8} = x - 8 + \\frac{64}{x - 8} + 16.\\]By AM-GM,\n\\[x - 8 + \\frac{64}{x - 8} \\ge 2 \\sqrt{(x - 8) \\cdot \\frac{64}{x - 8}} = 16,\\]so\n\\[\\frac{x^2}{x - 8} \\ge M.\\]Equality occurs when $x = 16,$ so the minimum value is $\\boxed{M}.$",
         "id": "train/intermediate_algebra/54.json",
-        "generated_solution": """
+        "generation": """
 To solve this problem we can
 - find critical points in the desired interval
 - evaluate the function at critical points
@@ -613,7 +768,7 @@ Thus the minimum value of the given function for $x > 8$ is $\\boxed{32}$.
         "reference_solution": "Since $-2$ and 3 are roots,\n\\begin{align*}\na(-2)^3 + (a + 2b) (-2)^2 + (b - 3a)(-2) + (8 - a) &= 0, \\\\\na(3)^3 + (a + 2b) 3^2 + (b - 3a)(3) + (8 - a) &= 0.\n\\end{align*}Solving, we find $a = \\frac{8}{9}$ and $b = -\\frac{40}{27}.$  By Vieta's formulas, the sum of the roots is\n\\[-\\frac{a + 2b}{a} = \\frac{7}{3},\\]so the third root is $\\frac{7}{3} - (-2) - 3 = \\boxed{\\frac{4}{3}}.$",
         "masked_reference_solution": "Since $-2$ and 3 are roots,\n\\begin{align*}\na(-2)^3 + (a + 2b) (-2)^2 + (b - 3a)(-2) + (8 - a) &= 0, \\\\\na(3)^3 + (a + 2b) 3^2 + (b - 3a)(3) + (8 - a) &= 0.\n\\end{align*}Solving, we find $a = \\frac{M}{N}$ and $b = -\\frac{O}{P}.$  By Vieta's formulas, the sum of the roots is\n\\[-\\frac{a + 2b}{a} = \\frac{Q}{R},\\]so the third root is $\\frac{Q}{R} - (-2) - 3 = \\boxed{\\frac{S}{T}}.$",
         "id": "train/intermediate_algebra/1516.json",
-        "generated_solution": """
+        "generation": """
 To find the third root of the given cubic equation
 $ax^3 + (a + 2b)x^2 + (b - 3a)x + (8 - a) = 0$,
 we first need to determine the values of the coefficients $a$ and $b$.
@@ -672,7 +827,7 @@ examples_number_theory = [
         "reference_solution": "According to the problem statement, we have the system of linear congruences \\begin{align*}\nm &\\equiv 0 \\pmod{6} \\\\\nm &\\equiv 2 \\pmod{8} \\\\\nm &\\equiv 2 \\pmod{5}.\n\\end{align*} It follows by the Chinese Remainder Theorem that $m \\equiv 2 \\pmod{40}$. The only number that satisfies this criterion for $30 \\le m \\le 80$ is $m = \\boxed{42}$, which is indeed divisible by $6$.",
         "masked_reference_solution": "According to the problem statement, we have the system of linear congruences \\begin{align*}\nm &\\equiv 0 \\pmod{6} \\\\\nm &\\equiv 2 \\pmod{8} \\\\\nm &\\equiv 2 \\pmod{5}.\n\\end{align*} It follows by the Chinese Remainder Theorem that $m \\equiv 2 \\pmod{M}$. The only number that satisfies this criterion for $30 \\le m \\le 80$ is $m = \\boxed{N}$, which is indeed divisible by $6$.",
         "id": "train/number_theory/77.json",
-        "generated_solution": """
+        "generation": """
 Since the interval is relatively small, we can check all the values from 30 to 80:
 {start_code}
 # Iterate through possible values of m between 30 and 80 (inclusive)
@@ -693,7 +848,7 @@ The only number that satisfies the requirements is $\\boxed{42}$.
         "reference_solution": "If the first three rows have 1 student, the last row must have two students, so there are 5 students in total. This is not greater than 30, so we must add another student to each row. This gives 9, which is still not greater than 30. We have to keep adding 4 until we get to a number greater than 30. As a result, we reach 13, 17, 21, 25, 29, 33. 33 is the first integer greater than 30, so this class has $\\boxed{33}$ students.",
         "masked_reference_solution": "If the first three rows have 1 student, the last row must have two students, so there are 5 students in total. This is not greater than 30, so we must add another student to each row. This gives 9, which is still not greater than 30. We have to keep adding 4 until we get to a number greater than 30. As a result, we reach 13, 17, 21, 25, 29, M. M is the first integer greater than 30, so this class has $\\boxed{M}$ students.",
         "id": "train/number_theory/119.json",
-        "generated_solution": """
+        "generation": """
 If $x$ is a number of students in one row, then total number of students is $3x + (x + 1)$.
 This should be greater or equals to 30: $3x + (x + 1) \\ge 30$
 {start_code}
@@ -731,7 +886,7 @@ The answer is $\\boxed{33}$.
         "reference_solution": "In order to solve this problem, we must first find the number of digits when $987_{10}$ is converted to each base. Beginning with base-3, we have that $2187>987>729$ or $3^7>987>3^6$. So, we know that the base-3 representation of $987_{10}$ has 7 digits. Similarly with base-8, we have that $4096>987>512$ or $8^4>987>8^3$. So, the base-8 representation of $987_{10}$ has only 4 digits. Therefore, the base-3 equivalent has $7-4=\\boxed{3}$ more digits than the base-8 equivalent.",
         "masked_reference_solution": "In order to solve this problem, we must first find the number of digits when $987_{10}$ is converted to each base. Beginning with base-3, we have that $2187>987>729$ or $3^7>987>3^6$. So, we know that the base-3 representation of $987_{10}$ has M digits. Similarly with base-8, we have that $4096>987>512$ or $8^4>987>8^3$. So, the base-8 representation of $987_{10}$ has only N digits. Therefore, the base-3 equivalent has $M-N=\\boxed{O}$ more digits than the base-8 equivalent.",
         "id": "train/number_theory/400.json",
-        "generated_solution": """
+        "generation": """
 We can easily do it in python.
 First, we need to create a function that changes base-10 numbers to other bases.
 Then we get base-3 and base-8 representations and calculate lengths difference.
@@ -766,7 +921,7 @@ So the answer is $\\boxed{3}$.
         "reference_solution": "We rewrite $AB$ as $10A+B$ and $AAB$ as $100A+10A+B$. Now we set $AAB=9\\cdot AB$ since $AB$ is $\\frac{1}{9}$ of $AAB$. \\begin{align*}\n100A+10A+B&=9(10A+B)\\quad\\Rightarrow\\\\\n&=90A+9B\\quad\\Rightarrow\\\\\n20A&=8B\\quad\\Rightarrow\\\\\n5A&=2B\n\\end{align*}The smallest possible values for $A$ and $B$ such that $5A=2B$ are $A=2$ and $B=5$. So $AAB=\\boxed{225}$.",
         "masked_reference_solution": "We rewrite $AB$ as $10A+B$ and $AAB$ as $100A+10A+B$. Now we set $AAB=9\\cdot AB$ since $AB$ is $\\frac{1}{9}$ of $AAB$. \\begin{align*}\n100A+10A+B&=9(10A+B)\\quad\\Rightarrow\\\\\n&=90A+9B\\quad\\Rightarrow\\\\\n20A&=8B\\quad\\Rightarrow\\\\\n5A&=2B\n\\end{align*}The smallest possible values for $A$ and $B$ such that $5A=2B$ are $A=M$ and $B=N$. So $AAB=\\boxed{MMN}$.",
         "id": "train/number_theory/512.json",
-        "generated_solution": """
+        "generation": """
 Let's iterate over all possible values of $a$ and $b$ and check what is the smallest value that satisfies conditions.
 {start_code}
 for a in range(1, 10):
@@ -794,7 +949,7 @@ Thus the smallest such number is $\\boxed{225}$
         "reference_solution": "$$ 3240 = 2^3 \\cdot 3^4 \\cdot 5^1 $$A positive divisor of 3240 is a multiple of 3 when it has a prime factorization in the form $2^a \\cdot 3^b \\cdot 5^c$ where $0 \\le a \\le 3$, $1 \\le b \\le 4$, and $0 \\le c \\le 1$.  There are $4 \\cdot 4 \\cdot 2 = \\boxed{32}$ choices for $a$, $b$, and $c$, giving the number of positive divisors of 3240 that are multiples of 3.",
         "masked_reference_solution": "$$ 3240 = 2^3 \\cdot 3^4 \\cdot 5^1 $$A positive divisor of 3240 is a multiple of 3 when it has a prime factorization in the form $2^a \\cdot 3^b \\cdot 5^c$ where $0 \\le a \\le 3$, $1 \\le b \\le 4$, and $0 \\le c \\le 1$.  There are $M \\cdot N \\cdot O = \\boxed{P}$ choices for $a$, $b$, and $c$, giving the number of positive divisors of 3240 that are multiples of 3.",
         "id": "train/number_theory/1270.json",
-        "generated_solution": """
+        "generation": """
 We can list all divisors of 3240 and filter only multiples of 3.
 Let's use sympy for this:
 {start_code}
@@ -825,7 +980,7 @@ examples_geometry = [
         "type": "Geometry",
         "reference_solution": "The two angles opposite the equal sides of an isosceles triangle are congruent, so in this case, both are $40^\\circ$. Since the three angles of a triangle add up to $180^\\circ$, the third angle in this triangle is $(180-40-40)^\\circ = \\boxed{100}^\\circ$.",
         "masked_reference_solution": "The two angles opposite the equal sides of an isosceles triangle are congruent, so in this case, both are $40^\\circ$. Since the three angles of a triangle add up to $180^\\circ$, the third angle in this triangle is $(180-40-40)^\\circ = \\boxed{M}^\\circ$.",
-        "generated_solution": """
+        "generation": """
 In an isosceles triangle two angles opposite the equal sides are equal.
 Let A and B be these angles.
 We know that the sum of all angles is 180 degrees.
@@ -855,7 +1010,7 @@ Since we need to find the measure of the largest angle, the answer is $\\boxed{1
         "type": "Geometry",
         "reference_solution": "The coordinates of point $B$ are $(-x,6)$.  The sum of all four coordinates is $x+6+(-x)+6=\\boxed{12}$.",
         "masked_reference_solution": "The coordinates of point $B$ are $(-x,6)$.  The sum of all four coordinates is $x+6+(-x)+6=\\boxed{M}$.",
-        "generated_solution": """
+        "generation": """
 Reflection over the $y$-axis changes $x$ coordinate to $-x$ and keeps $y$ coordinate constant.
 We can calculate the sum of the four coordinate values using sympy library:
 {start_code}
@@ -882,7 +1037,7 @@ Thus the answer is $\\boxed{12}$.
         "type": "Geometry",
         "reference_solution": "A sphere with radius $r$ has volume $\\frac{4}{3}\\pi r^3$.  Thus, the snowballs with radius 2, 3, and 5 inches have volumes $\\frac{4}{3}\\pi(2^3)$, $\\frac{4}{3}\\pi(3^3)$, and $\\frac{4}{3}\\pi(5^3)$ cubic inches respectively.  The total volume of snow used is thus  \\begin{align*}\n\\frac{4}{3}\\pi(2^3)+\\frac{4}{3}\\pi(3^3)+\\frac{4}{3}\\pi(5^3)&=\\frac{4}{3}\\pi(2^3+3^3+5^3)\\\\\n&=\\frac{4}{3}\\pi(8+27+125)\\\\\n&=\\boxed{\\frac{640}{3}\\pi}.\\end{align*}",
         "masked_reference_solution": "A sphere with radius $r$ has volume $\\frac{4}{3}\\pi r^3$.  Thus, the snowballs with radius 2, 3, and 5 inches have volumes $\\frac{4}{3}\\pi(2^3)$, $\\frac{4}{3}\\pi(3^3)$, and $\\frac{4}{3}\\pi(5^3)$ cubic inches respectively.  The total volume of snow used is thus  \\begin{align*}\n\\frac{4}{3}\\pi(2^3)+\\frac{4}{3}\\pi(3^3)+\\frac{4}{3}\\pi(5^3)&=\\frac{4}{3}\\pi(2^3+3^3+5^3)\\\\\n&=\\frac{4}{3}\\pi(M+N+O)\\\\\n&=\\boxed{\\frac{P}{Q}\\pi}.\\end{align*}",
-        "generated_solution": """
+        "generation": """
 To calculate the total volume of a snowman we need to sum volumes of three snowballs.
 The volume of a sphere with radius $r$ is $\\frac{4}{3}\\pi r^3$.
 We can use sympy for the calculations:
@@ -915,7 +1070,7 @@ So the total volume is $\\boxed{\\frac{640}{3}\\pi}$.
         "type": "Geometry",
         "reference_solution": "The fly's journey traces out the three sides of triangle.  Because one side of this triangle is a diameter of the circular ceiling and the other two sides are chords of this circle, the triangle is a right triangle. If the radius of the circular room is 58 feet, then the diameter is $2 \\times 58 = 116$ feet. This is the hypotenuse of the right triangle. One of the legs is 80 feet, so the other leg must be equal to $\\sqrt{116^2 - 80^2} = \\sqrt{(13{,}456 - 6400)} = \\sqrt{7056} = 84$ feet. The total distance traveled by the fly is $116 + 84 + 80 = \\boxed{280}$ feet.",
         "masked_reference_solution": "The fly's journey traces out the three sides of triangle.  Because one side of this triangle is a diameter of the circular ceiling and the other two sides are chords of this circle, the triangle is a right triangle. If the radius of the circular room is 58 feet, then the diameter is $2 \\times 58 = M$ feet. This is the hypotenuse of the right triangle. One of the legs is 80 feet, so the other leg must be equal to $\\sqrt{M^2 - 80^2} = \\sqrt{N} = O$ feet. The total distance traveled by the fly is $M + O + 80 = \\boxed{P}$ feet.",
-        "generated_solution": """
+        "generation": """
 Let x be the unknown second part of the journey and d be the diameter of the circle.
 The journey forms a triangle with sides d, x, and 80.
 Since this triangle is inscribed in a circle and one of its sides is a diameter,
@@ -951,7 +1106,7 @@ So the total distance traveled by the fly is $\\boxed{280}$.
         "type": "Geometry",
         "reference_solution": "Let the dimensions of $P$ be $x$, $y$, and $z$. The sum of the lengths of the edges of $P$ is $4(x+y+z)$, and the surface area of $P$ is $2xy+2yz+2xz$, so \\[\nx+y+z=28 \\quad\\text{and}\\quad 2xy+2yz+2xz=384.\n\\] Each internal diagonal of $P$ is a diameter of the sphere, so \\begin{align*}\n(2r)^2&=(x^2+y^2+z^2)\\\\\n&=(x+y+z)^2-(2xy+2xz+2yz) \\\\\n&= 28^2-384\\\\& = 400.\n\\end{align*} So $2r = 20$ and  $r=\\boxed{10}$.\n\nNote: There are infinitely many positive solutions of the system $x+y+z=28$, $2xy+2yz+2xz=384$, so there are infinitely many non-congruent boxes meeting the given conditions, but each can be inscribed in a sphere of radius 10.",
         "masked_reference_solution": "Let the dimensions of $P$ be $x$, $y$, and $z$. The sum of the lengths of the edges of $P$ is $4(x+y+z)$, and the surface area of $P$ is $2xy+2yz+2xz$, so \\[\nx+y+z=M \\quad\\text{and}\\quad 2xy+2yz+2xz=384.\n\\] Each internal diagonal of $P$ is a diameter of the sphere, so \\begin{align*}\n(2r)^2&=(x^2+y^2+z^2)\\\\\n&=(x+y+z)^2-(2xy+2xz+2yz) \\\\\n&= M^2-384\\\\& = 400.\n\\end{align*} So $2r = N$ and  $r=\\boxed{O}$.\n\nNote: There are infinitely many positive solutions of the system $x+y+z=M$, $2xy+2yz+2xz=384$, so there are infinitely many non-congruent boxes meeting the given conditions, but each can be inscribed in a sphere of radius O.",
-        "generated_solution": """
+        "generation": """
 To solve this problem, we need to use the properties of a rectangular box and a sphere.
 The box is inscribed in the sphere, which means that the sphere's diameter is the diagonal of the box.
 Assuming $x$, $y$, and $z$ to be dimensions of a box, we can make three equations.
@@ -996,7 +1151,7 @@ examples_precalculus = [
         "type": "Precalculus",
         "reference_solution": "We have that $\\det (\\mathbf{M}^4) = (\\det \\mathbf{M})^4 = \\boxed{16}.$",
         "masked_reference_solution": "We have that $\\det (\\mathbf{M}^4) = (\\det \\mathbf{M})^4 = \\boxed{N}.$",
-        "generated_solution": '''
+        "generation": '''
 We know that det(AB) = det(A) * det(B).
 We will use this fact to compute $\\det (\\mathbf{M}^4)$ using Python code.
 {start_code}
@@ -1019,7 +1174,7 @@ So the answer is \\boxed{16}.
         "type": "Precalculus",
         "reference_solution": "We know that $\\sin^2 x + \\cos^2 x = 1.$  Substituting $\\sin x = 3 \\cos x,$ we get\n\\[9 \\cos^2 x + \\cos^2 x = 1,\\]so $10 \\cos^2 x = 1,$ or $\\cos^2 x = \\frac{1}{10}.$  Then\n\\[\\sin x \\cos x = (3 \\cos x)(\\cos x) = 3 \\cos^2 x = \\boxed{\\frac{3}{10}}.\\]",
         "masked_reference_solution": "We know that $\\sin^2 x + \\cos^2 x = 1.$  Substituting $\\sin x = 3 \\cos x,$ we get\n\\[M \\cos^2 x + \\cos^2 x = 1,\\]so $N \\cos^2 x = 1,$ or $\\cos^2 x = \\frac{1}{N}.$  Then\n\\[\\sin x \\cos x = (3 \\cos x)(\\cos x) = 3 \\cos^2 x = \\boxed{\\frac{3}{N}}.\\]",
-        "generated_solution": '''
+        "generation": '''
 We know that $\\sin^2 x + \\cos^2 x = 1.$
 Substituting $\\sin x = 3 \\cos x$ in the above equation, we get  $9 \\cos^2 x + \\cos^2 x = 1$
 Thus, $\\cos^2 x = \\frac{1}{10}$.
@@ -1034,7 +1189,7 @@ Hence, the answer is $\\boxed{\\frac{3}{10}}$.
         "type": "Precalculus",
         "reference_solution": "Let $\\mathbf{b} = \\begin{pmatrix} x \\\\ y \\\\ z \\end{pmatrix}.$  Then the equation $\\mathbf{a} \\cdot \\mathbf{b} = 11$ gives us $2x + y + 5z = 11.$  Also,\n\\[\\mathbf{a} \\times \\mathbf{b} = \\begin{pmatrix} 2 \\\\ 1 \\\\ 5 \\end{pmatrix} \\times \\begin{pmatrix} x \\\\ y \\\\ z \\end{pmatrix} = \\begin{pmatrix} -5y + z \\\\ 5x - 2z \\\\ -x + 2y \\end{pmatrix}.\\]Comparing entries, we obtain\n\\begin{align*}\n-5y + z &= -13, \\\\\n5x - 2z &= -9, \\\\\n-x + 2y &= 7.\n\\end{align*}Solving this system, along with the equation $2x + y + z = 5z = 11,$ we find $x = -1,$ $y = 3,$ and $z = 2.$  Hence, $\\mathbf{b} = \\boxed{\\begin{pmatrix} -1 \\\\ 3 \\\\ 2 \\end{pmatrix}}.$",
         "masked_reference_solution": "Let $\\mathbf{b} = \\begin{pmatrix} x \\\\ y \\\\ z \\end{pmatrix}.$  Then the equation $\\mathbf{a} \\cdot \\mathbf{b} = 11$ gives us $M_1 * x + M_2 * y + M_3 * z = 11.$  Also,\n\\[\\mathbf{a} \\times \\mathbf{b} = \\begin{pmatrix} 2 \\\\ 1 \\\\ 5 \\end{pmatrix} \\times \\begin{pmatrix} x \\\\ y \\\\ z \\end{pmatrix} = \\begin{pmatrix} N_01 * y + N_02 * z \\\\ N_11 * x + N12 * z \\\\ N_21 * x + N_22 * y \\end{pmatrix}.\\]Comparing entries, we obtain\n\\begin{align*}\nN_01 * y + N_02 * z &= -13, \\\\\n N_11 * x + N12 * z &= -9, \\\\\n N_21 * x + N_22 * y &= 7.\n\\end{align*}Solving this system, along with the equation $M_1 * x + M_2 * y + M_3 * z = 11$ we find $x = O_0,$ $y = O_1,$ and $z = O_2.$  Hence, $\\mathbf{b} = \\boxed{\\begin{pmatrix} O_0 \\\\ O_1 \\\\ O_2 \\end{pmatrix}}.$",
-        "generated_solution": '''
+        "generation": '''
 Using the dot product and cross product constraints, we can solve the problem using the sympy library.
 {start_code}
 from sympy import symbols, Eq, solve
@@ -1074,7 +1229,7 @@ So the answer is $\\boxed{\\begin{pmatrix} -1 \\\\ 3 \\\\ 2 \\end{pmatrix}}$.
         "type": "Precalculus",
         "reference_solution": "We can write\n\\[\\tan \\frac{\\pi}{24} + \\tan \\frac{7 \\pi}{24} = \\frac{\\sin \\frac{\\pi}{24}}{\\cos \\frac{\\pi}{24}} + \\frac{\\sin \\frac{7 \\pi}{24}}{\\cos \\frac{7 \\pi}{24}} \n= \\frac{\\sin \\frac{\\pi}{24} \\cos \\frac{7 \\pi}{24} + \\cos \\frac{\\pi}{24} \\sin \\frac{7 \\pi}{24}}{\\cos \\frac{\\pi}{24} \\cos \\frac{7 \\pi}{24}}.\\]By the angle addition formula and the product-to-sum formula,\n\\begin{align*}\n\\frac{\\sin \\frac{\\pi}{24} \\cos \\frac{7 \\pi}{24} + \\cos \\frac{\\pi}{24} \\sin \\frac{7 \\pi}{24}}{\\cos \\frac{\\pi}{24} \\cos \\frac{7 \\pi}{24}} &= \\frac{\\sin (\\frac{\\pi}{24} + \\frac{7 \\pi}{24})}{\\frac{1}{2} (\\cos \\frac{\\pi}{3} + \\cos \\frac{\\pi}{4})} \\\\\n&= \\frac{2 \\sin \\frac{\\pi}{3}}{\\cos \\frac{\\pi}{3} + \\cos \\frac{\\pi}{4}} \\\\\n&= \\frac{\\sqrt{3}}{\\frac{1}{2} + \\frac{\\sqrt{2}}{2}} \\\\\n&= \\frac{2 \\sqrt{3}}{1 + \\sqrt{2}} \\\\\n&= \\frac{2 \\sqrt{3} (\\sqrt{2} - 1)}{(\\sqrt{2} + 1)(\\sqrt{2} - 1)} \\\\\n&= \\boxed{2 \\sqrt{6} - 2 \\sqrt{3}}.\n\\end{align*}",
         "masked_reference_solution": "We can write\n\\[\\tan \\frac{\\pi}{24} + \\tan \\frac{7 \\pi}{24} = \\frac{\\sin \\frac{\\pi}{24}}{\\cos \\frac{\\pi}{24}} + \\frac{\\sin \\frac{7 \\pi}{24}}{\\cos \\frac{7 \\pi}{24}} \n= \\frac{\\sin \\frac{\\pi}{24} \\cos \\frac{7 \\pi}{24} + \\cos \\frac{\\pi}{24} \\sin \\frac{7 \\pi}{24}}{\\cos \\frac{\\pi}{24} \\cos \\frac{7 \\pi}{24}}.\\]By the angle addition formula and the product-to-sum formula,\n\\begin{align*}\n\\frac{\\sin \\frac{\\pi}{24} \\cos \\frac{7 \\pi}{24} + \\cos \\frac{\\pi}{24} \\sin \\frac{7 \\pi}{24}}{\\cos \\frac{\\pi}{24} \\cos \\frac{7 \\pi}{24}} &= \\frac{\\sin (\\frac{\\pi}{24} + \\frac{7 \\pi}{24})}{\\frac{1}{2} (\\cos \\frac{\\pi}{3} + \\cos \\frac{\\pi}{4})} \\\\\n&= \\frac{2 \\sin \\frac{\\pi}{3}}{\\cos \\frac{\\pi}{3} + \\cos \\frac{\\pi}{4}} \\\\\n&= \\frac{\\sqrt{3}}{\\frac{1}{2} + \\frac{\\sqrt{2}}{2}} \\\\\n&= \\frac{2 \\sqrt{3}}{1 + \\sqrt{2}} \\\\\n&= \\frac{2 \\sqrt{3} (\\sqrt{2} - 1)}{(\\sqrt{2} + 1)(\\sqrt{2} - 1)} \\\\\n&= \\boxed{M}.\n\\end{align*}",
-        "generated_solution": '''
+        "generation": '''
 We will use sympy's simplify function to simplify the given trigonometric expression.
 {start_code}
 from sympy import tan, pi, simplify
@@ -1103,7 +1258,7 @@ So the answer is $\\boxed{2 \\sqrt{6} - 2 \\sqrt{3}}$.
         "type": "Precalculus",
         "reference_solution": "The side length of the regular tetrahedron is the distance between $(0,1,2)$ and $(4,2,1),$ which is\n\\[\\sqrt{(0 - 4)^2 + (1 - 2)^2 + (2 - 1)^2} = \\sqrt{18} = 3 \\sqrt{2}.\\]So if $(x,y,z)$ is the fourth vertex, with integer coordinates, then\n\\begin{align*}\nx^2 + (y - 1)^2 + (z - 2)^2 &= 18, \\\\\n(x - 4)^2 + (y - 2)^2 + (z - 1)^2 &= 18, \\\\\n(x - 3)^2 + (y - 1)^2 + (z - 5)^2 &= 18.\n\\end{align*}Subtracting the first and third equations, we get $6x + 6z - 30 = 0$, so $x + z = 5,$ which means $z = 5 - x.$  Subtracting the first and second equation, we get $8x + 2y - 2z - 16 = 0,$ so\n\\[y = z - 4x + 8 = (5 - x) - 4x + 8 = 13 - 5x.\\]Substituting into the first equation, we get\n\\[x^2 + (12 - 5x)^2 + (3 - x)^2 = 18.\\]This simplifies to $27x^2 - 126x + 135 = 0,$ which factors as $9(x - 3)(3x - 5) = 0.$  Since $x$ is an integer, $x = 3.$  Then $y = -2$ and $z = 2.$  Thus, the fourth vertex is $\\boxed{(3,-2,2)}.$",
         "masked_reference_solution": "The side length of the regular tetrahedron is the distance between $(0,1,2)$ and $(4,2,1),$ which is\n\\[\\sqrt{(0 - 4)^2 + (1 - 2)^2 + (2 - 1)^2} = \\sqrt{M}.\\]So if $(x,y,z)$ is the fourth vertex, with integer coordinates, then\n\\begin{align*}\nx^2 + (y - 1)^2 + (z - 2)^2 &= M, \\\\\n(x - 4)^2 + (y - 2)^2 + (z - 1)^2 &= M, \\\\\n(x - 3)^2 + (y - 1)^2 + (z - 5)^2 &= M.\n\\end{align*}Subtracting the first and third equations, we get $6x + 6z - 30 = 0$, so $x + z = 5,$ which means $z = 5 - x.$  Subtracting the first and second equation, we get $8x + 2y - 2z - 16 = 0,$ so\n\\[y = z - 4x + 8 = (5 - x) - 4x + 8 = 13 - 5x.\\]Substituting into the first equation, we get\n\\[x^2 + (12 - 5x)^2 + (3 - x)^2 = 18.\\]This simplifies to $Nx^2 - Ox + P = 0,$ which factors as $Q(x - R)(Sx - T) = 0.$  Since $x$ is an integer, $x = U$  Then $y = V$ and $z = W$  Thus, the fourth vertex is $\\boxed{(U,W,W)}.$",
-        "generated_solution": '''
+        "generation": '''
 The key insight to finding the missing vertex is that a regular tetrahedron has all the sides of equal length.
 Using the three provided vertices, we can create three equations constraining the fourth vertex and solve them to get solutions.
 Here' the solution in Python's sympy library.
@@ -1153,7 +1308,7 @@ examples_prealgebra = [
         "type": "Prealgebra",
         "reference_solution": "If Aaron ate the most apples, then we look at the highest column, which marks 6 apples eaten. Zeb ate the fewest apples, so we look for the shortest column, which marks 1 apple eaten. That means Aaron ate $6-1=\\boxed{5}$ more apples than Zeb.",
         "masked_reference_solution": "If Aaron ate the most apples, then we look at the highest column, which marks M apples eaten. Zeb ate the fewest apples, so we look for the shortest column, which marks N apple eaten. That means Aaron ate $M-N=\\boxed{O}$ more apples than Zeb.",
-        "generated_solution": """
+        "generation": """
 There are a total of six columns in the graph with values: 2, 3, 6, 4, 3, and 1.
 Since Aaron ate the most apples, and Zeb ate the fewest apples, we can answer the question using the following Python code.
 {start_code}
@@ -1180,7 +1335,7 @@ So Aaron ate \\boxed{5} more apples than Zeb.
         "type": "Prealgebra",
         "reference_solution": "Since the ratio of red to blue to green marbles is $1:5:3$, the ratio of green marbles to the total number of marbles is $3/(1+5+3) = 3/9=1/3$.  Since the green marbles are one-third of the total, and there are 27 green marbles, there must be $3\\cdot 27 = \\boxed{81}$ total marbles.",
         "masked_reference_solution": "Since the ratio of red to blue to green marbles is $1:5:3$, the ratio of green marbles to the total number of marbles is $3/(1+5+3) = 3/9=1/M$.  Since the green marbles are one-third of the total, and there are 27 green marbles, there must be $M\\cdot 27 = \\boxed{N}$ total marbles.",
-        "generated_solution": """
+        "generation": """
 To compute the total number of marbles, we can use Python's sympy library.
 {start_code}
 from sympy import Symbol, Eq, solve
@@ -1212,7 +1367,7 @@ Hence, there are a total of \\boxed{81} marbles.
         "type": "Prealgebra",
         "reference_solution": "The smallest perfect cube inside this range is $5^3 = 125$ since $4^3 = 64.$  As for the highest cube, we know that $10^3 = 1000,$ which is outside the range, so we try $9^3 = 729.$  Therefore the cubes in question are $5^3,6^3,7^3,8^3,9^3$.  So there are $\\boxed{5}$ such cubes.",
         "masked_reference_solution": "The smallest perfect cube inside this range is $5^3 = 125$ since $4^3 = 64.$  As for the highest cube, we know that $10^3 = 1000,$ which is outside the range, so we try $9^3 = 729.$  Therefore the cubes in question are $5^3,6^3,7^3,8^3,9^3$.  So there are $\\boxed{M}$ such cubes.",
-        "generated_solution": """
+        "generation": """
 To calculate these perfect cubes between 100 and 900, there are a few ways.
 One way is to enumerate cubes, and calculate the ones which satisfy this constraint.
 Another is to just take cube roots of the interval endpoints, with ceil and floor fns applied to start and end of the boundary.
@@ -1238,7 +1393,7 @@ Hence, there are a total of \\boxed{5} perfect cubes between 100 and 900.
         "type": "Prealgebra",
         "reference_solution": "We observe that the prime factorization of $120$ is equal to $2^3 \\cdot 3 \\cdot 5$.  It is a relatively quick matter to test that $2$, $3$, $4$, $5$, and $6$ share a prime factor with $120$, but $\\boxed{7}$ does not.",
         "masked_reference_solution": "We observe that the prime factorization of $120$ is equal to $2^3 \\cdot 3 \\cdot 5$.  It is a relatively quick matter to test that $2$, $3$, $4$, $5$, and $6$ share a prime factor with $120$, but $\\boxed{M}$ does not.",
-        "generated_solution": '''
+        "generation": '''
 To compute the smallest positive integer which is relatively prime to 120, we can iterate over all the numbers greater than 1 and less than 120, and break whenever we find a number which has a gcd of 1 with 120.
 {start_code}
 from sympy import gcd
@@ -1264,7 +1419,7 @@ Hence, \\boxed{7} is the smallest number which is relatively prime to 120.
         "type": "Prealgebra",
         "reference_solution": "Let each side of the square have length $x$. Then $AM=MB=BN=x/2$. So the area of the triangle is $(x/2)(x/2)/2=x^2/8$. The area of the square is $x\\cdot x= x^2$. The ratio of the two areas is $(x^2/8)/x^2=\\boxed{\\frac{1}{8}}$.",
         "masked_reference_solution": "Let each side of the square have length $x$. Then $AM=MB=BN=x/2$. So the area of the triangle is $(x/2)(x/2)/2=x^2/M$. The area of the square is $x\\cdot x= x^2$. The ratio of the two areas is $(x^2/M)/x^2=\\boxed{\\frac{1}{M}}$.",
-        "generated_solution": '''
+        "generation": '''
 Let each side of the square be 1. So the area of ABCD is 1^2=1.
 The triangle AMN, has base of length AM=1/2, and height of length of 1/2 as well since BN=1/2.
 Hence, area of AMN = (1/2) * base * height = (1/2) * (1/2) * (1/2) = 1/8.
@@ -1280,41 +1435,43 @@ generate_masked = [
         "level": "Level 1",
         "type": "Number Theory",
         "reference_solution": "According to the problem statement, we have the system of linear congruences \\begin{align*}\nm &\\equiv 0 \\pmod{6} \\\\\nm &\\equiv 2 \\pmod{8} \\\\\nm &\\equiv 2 \\pmod{5}.\n\\end{align*} It follows by the Chinese Remainder Theorem that $m \\equiv 2 \\pmod{40}$. The only number that satisfies this criterion for $30 \\le m \\le 80$ is $m = \\boxed{42}$, which is indeed divisible by $6$.",
-        "generated_solution": "According to the problem statement, we have the system of linear congruences \\begin{align*}\nm &\\equiv 0 \\pmod{6} \\\\\nm &\\equiv 2 \\pmod{8} \\\\\nm &\\equiv 2 \\pmod{5}.\n\\end{align*} It follows by the Chinese Remainder Theorem that $m \\equiv 2 \\pmod{M}$. The only number that satisfies this criterion for $30 \\le m \\le 80$ is $m = \\boxed{N}$, which is indeed divisible by $6$.",
+        "generation": "According to the problem statement, we have the system of linear congruences \\begin{align*}\nm &\\equiv 0 \\pmod{6} \\\\\nm &\\equiv 2 \\pmod{8} \\\\\nm &\\equiv 2 \\pmod{5}.\n\\end{align*} It follows by the Chinese Remainder Theorem that $m \\equiv 2 \\pmod{M}$. The only number that satisfies this criterion for $30 \\le m \\le 80$ is $m = \\boxed{N}$, which is indeed divisible by $6$.",
     },
     {
         "question": "I have a bag with only red, blue, and green marbles.  The ratio of red marbles to blue marbles to green marbles is $1:5:3$.  There are 27 green marbles in the bag.  How many marbles are there  in the bag?",
         "level": "Level 2",
         "type": "Prealgebra",
         "reference_solution": "Since the ratio of red to blue to green marbles is $1:5:3$, the ratio of green marbles to the total number of marbles is $3/(1+5+3) = 1/3$.  Since the green marbles are 1/3 of the total, and there are 27 green marbles, there must be $(3/1)\\cdot 27 = \\boxed{81}$ total marbles.",
-        "generated_solution": "Since the ratio of red to blue to green marbles is $1:5:3$, the ratio of green marbles to the total number of marbles is $3/(1+5+3) = M/N$.  Since the green marbles are M/N of the total, and there are 27 green marbles, there must be $(N/M)\\cdot 27 = \\boxed{O}$ total marbles.",
+        "generation": "Since the ratio of red to blue to green marbles is $1:5:3$, the ratio of green marbles to the total number of marbles is $3/(1+5+3) = M/N$.  Since the green marbles are M/N of the total, and there are 27 green marbles, there must be $(N/M)\\cdot 27 = \\boxed{O}$ total marbles.",
     },
     {
         "question": "Let $\\mathbf{a} = \\begin{pmatrix} 2 \\\\ 1 \\\\ 5 \\end{pmatrix}.$  Find the vector $\\mathbf{b}$ such that $\\mathbf{a} \\cdot \\mathbf{b} = 11$ and\n\\[\\mathbf{a} \\times \\mathbf{b} = \\begin{pmatrix} -13 \\\\ -9 \\\\ 7 \\end{pmatrix}.\\]",
         "level": "Level 3",
         "reference_solution": "Let $\\mathbf{b} = \\begin{pmatrix} x \\\\ y \\\\ z \\end{pmatrix}.$  Then the equation $\\mathbf{a} \\cdot \\mathbf{b} = 11$ gives us $2x + y + 5z = 11.$  Also,\n\\[\\mathbf{a} \\times \\mathbf{b} = \\begin{pmatrix} 2 \\\\ 1 \\\\ 5 \\end{pmatrix} \\times \\begin{pmatrix} x \\\\ y \\\\ z \\end{pmatrix} = \\begin{pmatrix} -5y + z \\\\ 5x - 2z \\\\ -x + 2y \\end{pmatrix}.\\]Comparing entries, we obtain\n\\begin{align*}\n-5y + z &= -13, \\\\\n5x - 2z &= -9, \\\\\n-x + 2y &= 7.\n\\end{align*}Solving this system, along with the equation $2x + y + z = 5z = 11,$ we find $x = -1,$ $y = 3,$ and $z = 2.$  Hence, $\\mathbf{b} = \\boxed{\\begin{pmatrix} -1 \\\\ 3 \\\\ 2 \\end{pmatrix}}.$",
-        "generated_solution": "Let $\\mathbf{b} = \\begin{pmatrix} x \\\\ y \\\\ z \\end{pmatrix}.$  Then the equation $\\mathbf{a} \\cdot \\mathbf{b} = 11$ gives us $M_1 * x + M_2 * y + M_3 * z = 11.$  Also,\n\\[\\mathbf{a} \\times \\mathbf{b} = \\begin{pmatrix} 2 \\\\ 1 \\\\ 5 \\end{pmatrix} \\times \\begin{pmatrix} x \\\\ y \\\\ z \\end{pmatrix} = \\begin{pmatrix} N_01 * y + N_02 * z \\\\ N_11 * x + N12 * z \\\\ N_21 * x + N_22 * y \\end{pmatrix}.\\]Comparing entries, we obtain\n\\begin{align*}\nN_01 * y + N_02 * z &= -13, \\\\\n N_11 * x + N12 * z &= -9, \\\\\n N_21 * x + N_22 * y &= 7.\n\\end{align*}Solving this system, along with the equation $M_1 * x + M_2 * y + M_3 * z = 11$ we find $x = O_0,$ $y = O_1,$ and $z = O_2.$  Hence, $\\mathbf{b} = \\boxed{\\begin{pmatrix} O_0 \\\\ O_1 \\\\ O_2 \\end{pmatrix}}.$",
+        "generation": "Let $\\mathbf{b} = \\begin{pmatrix} x \\\\ y \\\\ z \\end{pmatrix}.$  Then the equation $\\mathbf{a} \\cdot \\mathbf{b} = 11$ gives us $M_1 * x + M_2 * y + M_3 * z = 11.$  Also,\n\\[\\mathbf{a} \\times \\mathbf{b} = \\begin{pmatrix} 2 \\\\ 1 \\\\ 5 \\end{pmatrix} \\times \\begin{pmatrix} x \\\\ y \\\\ z \\end{pmatrix} = \\begin{pmatrix} N_01 * y + N_02 * z \\\\ N_11 * x + N12 * z \\\\ N_21 * x + N_22 * y \\end{pmatrix}.\\]Comparing entries, we obtain\n\\begin{align*}\nN_01 * y + N_02 * z &= -13, \\\\\n N_11 * x + N12 * z &= -9, \\\\\n N_21 * x + N_22 * y &= 7.\n\\end{align*}Solving this system, along with the equation $M_1 * x + M_2 * y + M_3 * z = 11$ we find $x = O_0,$ $y = O_1,$ and $z = O_2.$  Hence, $\\mathbf{b} = \\boxed{\\begin{pmatrix} O_0 \\\\ O_1 \\\\ O_2 \\end{pmatrix}}.$",
     },
     {
         "question": "Kadin makes a snowman by stacking snowballs of radius 2 inches, 3 inches, and 5 inches.  Assuming all his snowballs are spherical, what is the total volume of snow he uses, in cubic inches?  Express your answer in terms of $\\pi$.",
         "level": "Level 3",
         "type": "Geometry",
         "reference_solution": "A sphere with radius $r$ has volume $\\frac{4}{3}\\pi r^3$.  Thus, the snowballs with radius 2, 3, and 5 inches have volumes $\\frac{4}{3}\\pi(2^3)$, $\\frac{4}{3}\\pi(3^3)$, and $\\frac{4}{3}\\pi(5^3)$ cubic inches respectively.  The total volume of snow used is thus  \\begin{align*}\n\\frac{4}{3}\\pi(2^3)+\\frac{4}{3}\\pi(3^3)+\\frac{4}{3}\\pi(5^3)&=\\frac{4}{3}\\pi(2^3+3^3+5^3)\\\\\n&=\\frac{4}{3}\\pi(8+27+125)\\\\\n&=\\boxed{\\frac{640}{3}\\pi}.\\end{align*}",
-        "generated_solution": "A sphere with radius $r$ has volume $\\frac{4}{3}\\pi r^3$.  Thus, the snowballs with radius 2, 3, and 5 inches have volumes $\\frac{4}{3}\\pi(2^3)$, $\\frac{4}{3}\\pi(3^3)$, and $\\frac{4}{3}\\pi(5^3)$ cubic inches respectively.  The total volume of snow used is thus  \\begin{align*}\n\\frac{4}{3}\\pi(2^3)+\\frac{4}{3}\\pi(3^3)+\\frac{4}{3}\\pi(5^3)&=\\frac{4}{3}\\pi(2^3+3^3+5^3)\\\\\n&=\\frac{4}{3}\\pi(M+N+O)\\\\\n&=\\boxed{\\frac{P}{Q}\\pi}.\\end{align*}",
+        "generation": "A sphere with radius $r$ has volume $\\frac{4}{3}\\pi r^3$.  Thus, the snowballs with radius 2, 3, and 5 inches have volumes $\\frac{4}{3}\\pi(2^3)$, $\\frac{4}{3}\\pi(3^3)$, and $\\frac{4}{3}\\pi(5^3)$ cubic inches respectively.  The total volume of snow used is thus  \\begin{align*}\n\\frac{4}{3}\\pi(2^3)+\\frac{4}{3}\\pi(3^3)+\\frac{4}{3}\\pi(5^3)&=\\frac{4}{3}\\pi(2^3+3^3+5^3)\\\\\n&=\\frac{4}{3}\\pi(M+N+O)\\\\\n&=\\boxed{\\frac{P}{Q}\\pi}.\\end{align*}",
     },
     {
         "question": "In square $ABCD$, point $M$ is the midpoint of side $AB$ and point $N$ is the midpoint of side $BC$. What is the ratio of the area of triangle $AMN$ to the area of square $ABCD$? Express your answer as a common fraction.",
         "level": "Level 5",
         "type": "Prealgebra",
-        "generated_solution": "Let each side of the square have length $x$. Then $AM=MB=BN=x/2$. So the area of the triangle is (base * height)/2=x^2/O. The area of the square is $x\\cdot x= x^2$. The ratio of the two areas is $(x^2/O)/x^2=\\boxed{\\frac{1}{O}}$.",
+        "generation": "Let each side of the square have length $x$. Then $AM=MB=BN=x/2$. So the area of the triangle is (base * height)/2=x^2/O. The area of the square is $x\\cdot x= x^2$. The ratio of the two areas is $(x^2/O)/x^2=\\boxed{\\frac{1}{O}}$.",
         "reference_solution": "Let each side of the square have length $x$. Then $AM=MB=BN=x/2$. So the area of the triangle is $(x/2)(x/2)/2=x^2/8$. The area of the square is $x\\cdot x= x^2$. The ratio of the two areas is $(x^2/8)/x^2=\\boxed{\\frac{1}{8}}$.",
     },
 ]
 
 
 examples_map = {
+    "math_standard_few_shot": standard_four_shot,
     "math_text_with_code": text_with_code,
     "math_generate_masked": generate_masked,
+    "math_text_detailed": text_detailed,
     # 7 subjects
     "math_algebra": examples_algebra,
     "math_probability": examples_probability,
