@@ -37,6 +37,73 @@ def fill_up_missing(evaluations, log_probs, pred_answers, allow_incomplete):
         raise RuntimeError("Need to run evaluate_results.py before computing metrics!")
 
 
+def compute_metrics(prediction_jsonl_files, allow_incomplete=False, max_samples=-1, aggregation_mode='first'):
+    correct_answer = []
+
+    file_handles = [open(file, "rt", encoding="utf-8") for file in unroll_files(prediction_jsonl_files)]
+
+    total_correct = 0
+    total_no_answer = 0
+    total = 0
+    for idx, lines in enumerate(zip_longest(*file_handles)):
+        if idx == max_samples:
+            break
+        evaluations = []
+        log_probs = []
+        pred_answers = []
+        for lidx, line in enumerate(lines):
+            if not line:  # could have missing predictions
+                fill_up_missing(evaluations, log_probs, pred_answers, allow_incomplete)
+                continue
+            line_dict = json.loads(line)
+            if not line_dict:
+                fill_up_missing(evaluations, log_probs, pred_answers, allow_incomplete)
+                continue
+            if "is_correct" not in line_dict:
+                fill_up_missing(evaluations, log_probs, pred_answers, allow_incomplete)
+                continue
+            pred_answers.append(line_dict["predicted_answer"])
+            evaluations.append(line_dict["is_correct"])
+            log_probs.append(line_dict.get("log_prob"))
+            if line_dict.get("log_prob") is None and aggregation_mode == "most_probable":
+                raise RuntimeError(
+                    "Cannot compute most probable generation because some of the probabilities are missing"
+                )
+
+        total += 1
+        if aggregation_mode == "best":
+            total_correct += any(evaluations)
+            if all([ans is None for ans in pred_answers]):
+                total_no_answer += 1
+        elif aggregation_mode == "majority":
+            # TODO: currently majority does not take into account equivalent answers written in a different way
+            valid_answers_and_results = [
+                (ans, is_correct) for ans, is_correct in zip(pred_answers, evaluations) if ans is not None
+            ]
+            if len(valid_answers_and_results) == 0:
+                total_no_answer += 1
+            else:
+                majority_result = Counter(valid_answers_and_results).most_common(1)[0][0]
+                total_correct += majority_result[1]
+        elif aggregation_mode == "first":
+            total_correct += evaluations[0]
+            total_no_answer += pred_answers[0] is None
+        elif aggregation_mode == "most_probable":
+            most_probable_result = sorted(zip(log_probs, evaluations, pred_answers))[-1]
+            total_correct += most_probable_result[1]
+            total_no_answer += most_probable_result[2] is None
+        else:
+            raise ValueError(f"Unsupported mode {aggregation_mode}")
+
+    for file_handle in file_handles:
+        file_handle.close()
+
+    correct_answer = total_correct / total * 100.0
+    no_answer = total_no_answer / total * 100.0
+    wrong_answer = (total - total_correct - total_no_answer) / total * 100.0
+    return correct_answer, wrong_answer, no_answer, total
+
+
 if __name__ == '__main__':
     setup_logging(disable_hydra_logs=False)
     parser = argparse.ArgumentParser()
@@ -69,69 +136,10 @@ if __name__ == '__main__':
     )
     args = parser.parse_args()
 
-    correct_answer = []
+    correct_answer, wrong_answer, no_answer, total = compute_metrics(
+        args.prediction_jsonl_files, args.allow_incomplete, args.max_samples, args.aggregation_mode
+    )
 
-    file_handles = [open(file, "rt", encoding="utf-8") for file in unroll_files(args.prediction_jsonl_files)]
-
-    total_correct = 0
-    total_no_answer = 0
-    total = 0
-    for idx, lines in enumerate(zip_longest(*file_handles)):
-        if idx == args.max_samples:
-            break
-        evaluations = []
-        log_probs = []
-        pred_answers = []
-        for lidx, line in enumerate(lines):
-            if not line:  # could have missing predictions
-                fill_up_missing(evaluations, log_probs, pred_answers, args.allow_incomplete)
-                continue
-            line_dict = json.loads(line)
-            if not line_dict:
-                fill_up_missing(evaluations, log_probs, pred_answers, args.allow_incomplete)
-                continue
-            if "is_correct" not in line_dict:
-                fill_up_missing(evaluations, log_probs, pred_answers, args.allow_incomplete)
-                continue
-            pred_answers.append(line_dict["predicted_answer"])
-            evaluations.append(line_dict["is_correct"])
-            log_probs.append(line_dict.get("log_prob"))
-            if line_dict.get("log_prob") is None and args.aggregation_mode == "most_probable":
-                raise RuntimeError(
-                    "Cannot compute most probable generation because some of the probabilities are missing"
-                )
-
-        total += 1
-        if args.aggregation_mode == "best":
-            total_correct += any(evaluations)
-            if all([ans is None for ans in pred_answers]):
-                total_no_answer += 1
-        elif args.aggregation_mode == "majority":
-            # TODO: currently majority does not take into account equivalent answers written in a different way
-            valid_answers_and_results = [
-                (ans, is_correct) for ans, is_correct in zip(pred_answers, evaluations) if ans is not None
-            ]
-            if len(valid_answers_and_results) == 0:
-                total_no_answer += 1
-            else:
-                majority_result = Counter(valid_answers_and_results).most_common(1)[0][0]
-                total_correct += majority_result[1]
-        elif args.aggregation_mode == "first":
-            total_correct += evaluations[0]
-            total_no_answer += pred_answers[0] is None
-        elif args.aggregation_mode == "most_probable":
-            most_probable_result = sorted(zip(log_probs, evaluations, pred_answers))[-1]
-            total_correct += most_probable_result[1]
-            total_no_answer += most_probable_result[2] is None
-        else:
-            raise ValueError(f"Unsupported mode {args.aggregation_mode}")
-
-    for file_handle in file_handles:
-        file_handle.close()
-
-    correct_answer = total_correct / total * 100.0
-    no_answer = total_no_answer / total * 100.0
-    wrong_answer = (total - total_correct - total_no_answer) / total * 100.0
     LOG.info(f"Evaluation results for %s", args.prediction_jsonl_files)
     LOG.info(f"Total eval entries: %d", total)
     LOG.info(f"Correct answer: %.2f%%", correct_answer)
