@@ -12,6 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import os
 import sys
 from argparse import ArgumentParser
 from pathlib import Path
@@ -71,23 +72,23 @@ BENCHMARKS = {
     "tabmwp": 4,
 }
 
-# TODO: remove backoff installation after it gets into docker
 
 SLURM_CMD = """
 nvidia-smi && \
 cd /code && \
 export PYTHONPATH=$PYTHONPATH:/code && \
-{server_start_cmd} && \
-if [ $SLURM_LOCALID -eq 0 ]; then \
-    pip install backoff && \
+export HF_TOKEN={HF_TOKEN} && \
+if [ $SLURM_PROCID -eq 0 ]; then \
+    {{ {server_start_cmd} 2>&1 | tee /tmp/server_logs.txt & }} && sleep 1 && \
     echo "Waiting for the server to start" && \
     tail -n0 -f /tmp/server_logs.txt | sed '/{server_wait_string}/ q' && \
     {eval_cmds} \
-    kill %1 || true; \
+    pkill -f nemo_skills/inference/server; \
 else \
-    sleep infinity; \
+    {server_start_cmd}; \
 fi \
 """
+
 
 MOUNTS = "{NEMO_SKILLS_CODE}:/code,{model_path}:/model,{output_dir}:/results"
 JOB_NAME = "eval-{model_name}"
@@ -111,8 +112,15 @@ if __name__ == "__main__":
     wrapper_args.add_argument(
         "--num_nodes",
         type=int,
+        default=1,
+        help="Number of nodes required for hosting LLM server.",
+    )
+    wrapper_args.add_argument(
+        "--num_jobs",
+        type=int,
         default=-1,
-        help="Will parallelize across this number of nodes. Set -1 to run each decoding on a separate node.",
+        help="Will launch this many separate jobs and split the benchmarks across them. "
+        "Set -1 to run each benchmark / random seed as a separate job.",
     )
     wrapper_args.add_argument(
         "--partition",
@@ -133,7 +141,7 @@ if __name__ == "__main__":
     args.output_dir = Path(args.output_dir).absolute()
 
     server_start_cmd, num_tasks, server_wait_string = get_server_command(
-        args.server_type, args.num_gpus, args.model_path.name
+        args.server_type, args.num_gpus, args.num_nodes, args.model_path.name
     )
 
     format_dict = {
@@ -144,6 +152,7 @@ if __name__ == "__main__":
         "server_start_cmd": server_start_cmd,
         "server_type": args.server_type,
         "NEMO_SKILLS_CODE": NEMO_SKILLS_CODE,
+        "HF_TOKEN": os.getenv("HF_TOKEN", ""),  # needed for some of the models, so making an option to pass it in
         "server_wait_string": server_wait_string,
     }
 
@@ -162,17 +171,17 @@ if __name__ == "__main__":
         for benchmark, rs_num in BENCHMARKS.items()
         for rs in range(args.starting_seed, args.starting_seed + rs_num)
     ]
-    if args.num_nodes == -1:
-        args.num_nodes = len(eval_cmds)
+    if args.num_jobs == -1:
+        args.num_jobs = len(eval_cmds)
 
-    # splitting eval cmds equally across num_nodes nodes
-    eval_cmds = [" ".join(eval_cmds[i :: args.num_nodes]) for i in range(args.num_nodes)]
+    # splitting eval cmds equally across num_jobs nodes
+    eval_cmds = [" ".join(eval_cmds[i :: args.num_jobs]) for i in range(args.num_jobs)]
 
     for idx, eval_cmd in enumerate(eval_cmds):
         extra_sbatch_args = ["--parsable", f"--output={args.output_dir}/slurm_logs_eval{idx}.log"]
         launch_job(
             cmd=SLURM_CMD.format(**format_dict, eval_cmds=eval_cmd.format(**format_dict)),
-            num_nodes=1,
+            num_nodes=args.num_nodes,
             tasks_per_node=num_tasks,
             gpus_per_node=format_dict["num_gpus"],
             job_name=JOB_NAME.format(**format_dict),
