@@ -13,6 +13,9 @@
 # limitations under the License.
 
 # a helper script to check that running --help works on all scripts
+# not launching via pytest since it needs a different environment
+# (no packages installed) than other tests
+
 import argparse
 import subprocess
 import sys
@@ -22,7 +25,7 @@ pipeline_script_list = [
     'pipeline/prepare_eval.py',
     'pipeline/run_eval.py',
     'pipeline/run_labeling.py',
-    'pipeline/run_sft_and_eval.py',
+    'pipeline/run_pipeline.py',
     'pipeline/run_sft.py',
     'pipeline/start_server.py',
     'pipeline/summarize_results.py',
